Update illustrations
Illustrations are now rendered in pure svg
