--- src/components/svg/HeroIllustration.jsx
@@ -1,420 +1,0 @@
-export default function HeroIllustration() {
- return (
-  <>
-   <svg
-    width="400"
-    height="300"
-    xmlns="http://www.w3.org/2000/svg"
-    xmlnsXlink="http://www.w3.org/1999/xlink"
-    viewBox="0 0 400 300"
-   >
-    <defs>
-     <linearGradient
-      id="a"
-      x1="92.37"
-      x2="119.91"
-      y1="159.78"
-      y2="130.94"
-      gradientUnits="userSpaceOnUse"
-     >
-      <stop offset="0" stopColor="#dd6f31" />
-      <stop offset="1" stopColor="#532a12" />
-     </linearGradient>
-     <linearGradient
-      id="b"
-      x1="201.12"
-      x2="225.84"
-      y1="217.43"
-      y2="143.26"
-      xlinkHref="#a"
-     />
-     <linearGradient
-      id="c"
-      x1="36.66"
-      x2="55.37"
-      y1="192.48"
-      y2="172.89"
-      xlinkHref="#a"
-     />
-    </defs>
-    <path
-     fill="#dd6f31"
-     d="M193.23 287.42L13.5 181.2c-5.72-3.2-10.63-8.49-5.05-11.95l180.37-104c2.71-1.68 6.82-2.78 9.57-1.18L387 173.61c5.72 3.32 6.75 12.6 1 15.93l-171.08 97.71a23.9 23.9 0 0 1-23.69.17z"
-    />
-    <path
-     fill="url(#a)"
-     d="M104.37 172.75c4.42-2.71 25.44-16.1 29.78-18.92C141.59 149 151.09 142 151.09 142l-47.39-22.27-39.76 26.11c7.2 3.77 28.91 29.79 32.51 31.8 1.61-.95 6.34-3.92 7.92-4.89z"
-    />
-    <path
-     fill="url(#b)"
-     d="M209.59 236.52c15-8.07 61.1-45.06 62-47.66L221 156.26l-73.43 38.81c15.67 8.21 40.35 28.04 62.02 41.45z"
-    />
-    <path
-     fill="#56403d"
-     d="M240 147.11l-19 9.15a4.38 4.38 0 0 0-.31 7.72l45.06 26.28a4.38 4.38 0 0 0 5.89-1.41l10.95-17"
-    />
-    <path
-     fill="#68e1fd"
-     d="M335.49 98.39L180.3 9a5 5 0 0 0-7.57 4.37v95a3.89 3.89 0 0 0 1.95 3.37l165.18 95.35a1.88 1.88 0 0 0 2.82-1.63v-94.64a14.35 14.35 0 0 0-7.19-12.43z"
-    />
-    <path
-     fill="#fff"
-     d="M183 17.39L330 102a10.31 10.31 0 0 1 5.17 8.93V179a3.72 3.72 0 0 1-5.57 3.23l-149.91-86a3.67 3.67 0 0 1-1.86-3.23V20.39a3.45 3.45 0 0 1 5.17-3z"
-    />
-    <ellipse
-     cx="246.42"
-     cy="145.54"
-     fill="#ab643c"
-     rx="1.6"
-     ry=".01"
-     transform="rotate(-10.9 246.509 145.574)"
-    />
-    <ellipse cx="324.01" cy="104.06" fill="#659fe4" rx="1.6" ry="1.74" />
-    <ellipse cx="318.91" cy="101.52" fill="#c5e9ff" rx="1.6" ry="1.74" />
-    <path
-     fill="#00a99d"
-     d="M184.72 34.95l35.42 18.94c1.7.91 3.22-1.68 1.51-2.59l-35.41-18.94c-1.7-.91-3.22 1.68-1.51 2.59z"
-    />
-    <path
-     fill="url(#c)"
-     d="M46.15 200.47q6.63-4.07 13.16-8.3c7.43-4.83 10.59-5.76 17.5-11.53l-29.65-13-30.4 15.48c7.2 3.77 25.8 15.35 29.39 17.35 1.61-.94-1.59.98 0 0z"
-    />
-    <path
-     fill="#0a1d23"
-     d="M61.92 126.43c-13.66 0-24.74 3.62-24.74 8.09h49.48c0-4.52-11.07-8.09-24.74-8.09z"
-    />
-    <path
-     fill="#06489e"
-     d="M37.26 135.14l-.08-.63a2.74 2.74 0 0 0 .08.63zM86.66 134.52l-.1.71a2.76 2.76 0 0 0 .1-.71z"
-    />
-    <path
-     fill="#0a1d23"
-     d="M61.92 142.6c12.93 0 23.54-3.24 24.64-7.38l.1-.71H37.18l.08.63c.98 4.17 11.64 7.46 24.66 7.46z"
-    />
-    <path fill="#25b000" d="M65.81 92.76h5.57v57.42h-5.57z" />
-    <ellipse cx="68.6" cy="92.76" fill="#25b000" rx="2.78" ry="1.18" />
-    <path
-     fill="#ff8f02"
-     d="M78.293 102.608l5.458 1.114-11.487 56.26-5.458-1.115z"
-    />
-    <ellipse
-     cx="81.02"
-     cy="103.16"
-     fill="#ff8f02"
-     rx="1.18"
-     ry="2.78"
-     transform="rotate(-78.46 81.018 103.156)"
-    />
-    <path
-     fill="#68e1fd"
-     d="M55.2 135.78l-5.08 2.91a.45.45 0 0 1-.45-.78l5.2-3a.46.46 0 0 1 .16-.06l-1-5.5-5.13 2.95a.45.45 0 0 1-.45-.78l5.2-3a.45.45 0 0 1 .19-.06l-.94-5.09-5 2.87a.45.45 0 0 1-.45-.78l5.2-3h.1l-1.05-5.67-5 2.86a.45.45 0 0 1-.45-.78l5.2-3h.08l-1-5.54-5.08 2.91a.45.45 0 0 1-.45-.78l5.2-3a.46.46 0 0 1 .16-.06l-.89-4.81-5.09 2.91a.45.45 0 0 1-.45-.78l5.2-3a.46.46 0 0 1 .17-.06l-.89-4.8-5.08 2.9a.45.45 0 0 1-.45-.78l5.2-3a.46.46 0 0 1 .12.02l-1-5.28-5 2.87a.45.45 0 0 1-.45-.78l5.2-3h.1l-.46-2.46-11 6.84s7.84 43.65 8 43.74c4.58 2.39 12.91 2.67 12.91 2.67z"
-    />
-    <path
-     fill="#ca4300"
-     d="M41.74 94.83a.45.45 0 0 0 .45.78l5-2.87-.19-.93h-.1zM42.84 101a.45.45 0 0 0 .45.78l5.08-2.9-.17-.88a.46.46 0 0 0-.15.05zM43.89 106.76a.45.45 0 0 0 .45.78l5.09-2.91-.17-.9a.46.46 0 0 0-.17.06zM45 112.46a.45.45 0 0 0 .45.78l5.08-2.91-.17-.9a.46.46 0 0 0-.16.06zM46.23 118.88a.45.45 0 0 0 .45.78l5-2.86-.17-.93h-.08zM47.43 125.49a.45.45 0 0 0 .45.78l5-2.87-.17-.92h-.1zM53.65 128.55l-5.2 3a.45.45 0 0 0 .45.78l5.1-2.95-.17-.89a.45.45 0 0 0-.18.06zM49.67 137.9a.45.45 0 0 0 .45.78l5.08-2.91-.17-.9a.46.46 0 0 0-.16.06z"
-    />
-    <path
-     fill="#ffd8bd"
-     d="M78.29 102.6s.28 1.46 2.49 1.71 3-.6 3-.6l-.72-11z"
-    />
-    <path fill="#653200" d="M83 92.76l-2.2 4.57a2 2 0 0 0 2.52.35z" />
-    <path
-     fill="#ffd8bd"
-     d="M65.81 92.42s.51 1.4 2.73 1.3 2.83-1.06 2.83-1.06L68.94 82z"
-    />
-    <path fill="#653200" d="M68.94 82l-1.45 4.86a2 2 0 0 0 2.51-.09z" />
-    <path
-     fill="#173743"
-     d="M61.92 142.6c-13 0-23.68-3.29-24.66-7.46.65 5 5.09 39.32 5.26 39.52 0 0 2.27 9.68 19.4 9.67 14.57 0 18.73-7.32 18.73-7.32l5.91-41.79c-1.1 4.14-11.71 7.38-24.64 7.38z"
-    />
-    <path
-     fill="#ffe8bc"
-     d="M107.3 53.16l-8.87 4.73 24.21 22.8 21.35 6.7 7.1-9.64"
-    />
-    <path
-     fill="#68e1fd"
-     d="M140.72 82.54l8.23-4.69-41.65-24.69.98-.57 42.81 25.16V142l-9.34 5.32-3.56-3.72"
-    />
-    <path
-     fill="#1371a5"
-     d="M140.72 82.54l8.23-4.69-41.65-24.69.98-.57 42.81 25.16V142l-9.34 5.32-3.56-3.72"
-     opacity=".42"
-    />
-    <path
-     fill="#68e1fd"
-     d="M98.43 57.89l43.32 25.25v64.18l-44-26.3.68-63.13z"
-    />
-    <path
-     fill="#e85858"
-     d="M140.12 78.93l-2.65 1.68v13.68l-1.68-2.15-1.18 1.33V78.98l2.48-1.78 3.03 1.73z"
-    />
-    <path
-     fill="#ffe8bc"
-     d="M95.18 71.59l-7.91 4.22 21.58 20.32 19.02 5.98 6.34-8.59"
-    />
-    <path
-     fill="#68e1fd"
-     d="M124.96 97.78l7.34-4.17-37.12-22.02.87-.5 38.16 22.43v57.26l-8.33 4.74-3.17-3.31"
-    />
-    <path
-     fill="#1371a5"
-     d="M124.96 97.78l7.34-4.17-37.12-22.02.87-.5 38.16 22.43v57.26l-8.33 4.74-3.17-3.31"
-     opacity=".42"
-    />
-    <path
-     fill="#68e1fd"
-     d="M87.27 75.81l38.61 22.51v57.2l-39.22-23.44.61-56.27z"
-    />
-    <path
-     fill="#e85858"
-     d="M124.42 94.57l-2.35 1.49v12.19l-1.5-1.91-1.05 1.18V94.61l2.2-1.58 2.7 1.54z"
-    />
-    <path
-     fill="#ffe0de"
-     d="M141.93 230s-.47 3.66.07 3.88a4.88 4.88 0 0 0 3.26.65 44.92 44.92 0 0 1 4.91-1.29c2-.29 7.18.32 8.07-1.06s0-4.25 0-4.25"
-    />
-    <path
-     fill="#56403d"
-     d="M158.23 227.95c-.4-2.23-3.7-1.3-5.25-2.31-2.85-1.84-3-5.89-2.74-8.92l-6.34-2.2-1.25 7.28-.72 8.24a3.42 3.42 0 0 0 2.83 1.44l4.23-.94c2.69-.25 9.52-1.03 9.24-2.59z"
-    />
-    <path
-     fill="#ffe0de"
-     d="M150.77 234.35s-.47 3.66.07 3.88a4.88 4.88 0 0 0 3.26.65 44.92 44.92 0 0 1 4.91-1.29c2-.29 7.18.32 8.07-1.06s0-4.25 0-4.25"
-    />
-    <path
-     fill="#56403d"
-     d="M167.07 232.26c-.4-2.23-3.7-1.3-5.25-2.31-2.85-1.84-3-5.89-2.74-8.92l-6.34-2.2-1.25 7.28-.72 8.24a3.42 3.42 0 0 0 2.83 1.44l4.23-.94c2.69-.25 9.52-1.02 9.24-2.59z"
-    />
-    <path
-     fill="#e85b5b"
-     d="M141.61 210l-1.61 8.55s2.46 2.53 10.75 2.23 0-15.14 0-15.14"
-    />
-    <ellipse cx="130.85" cy="229.95" fill="#56403d" rx="11.8" ry="4.38" />
-    <path fill="#56403d" d="M128.86 210.34l.46 17.65h3.52v-17.65" />
-    <path
-     fill="#008074"
-     d="M153.54 184.87q-5.3 1.16-10.69 1.79a.6.6 0 0 0-.41.22q-.62-1.86-1.06-3.79c-.16-.73-1.28-.42-1.12.31a40.46 40.46 0 0 0 2.93 8.45c.32.68 1.32.09 1-.59-.53-1.13-1-2.27-1.43-3.44h.08q5.54-.65 11-1.84c.74-.15.43-1.27-.3-1.11zM144 193.06a15.6 15.6 0 0 1-3.58 4.11c-.58.47.25 1.29.82.82a16.5 16.5 0 0 0 3.76-4.34.58.58 0 0 0-1-.59z"
-    />
-    <path
-     fill="#edbab9"
-     d="M136.76 189.28s9.25-7.42 15.39-8.3 9.35 5.79 8.91 10.25-15.14 6.9-15.14 6.9"
-    />
-    <path
-     fill="#ff7e7b"
-     d="M141.69 191.75c1.2-1.11 16.9-8.44 22.43-5.09s5.18 5.82 4.48 9.9-6 26-6 26a20 20 0 0 1-12.56.59l1.55-16.31-20.92-11.42"
-    />
-    <path
-     fill="#fcca4c"
-     d="M130.25 165.51c.83.61 8 1.47 9.5 4.78s3.78 20.34 3.78 20.34-4.3 6.41-11.75 4.13c-2.21-.68-13.67-21.63-14-23.8-.78-5.66 12.47-5.45 12.47-5.45z"
-    />
-    <path
-     fill="#68e1fd"
-     d="M116.11 202.32a18.76 18.76 0 0 1-.89-2.65c-1.76-6.89-2.91-32.76 8.72-31.7 12.76 1.15 8.45 33.1 21.38 32.12s18.23 2.38 12.94 7.39c-10.57 10.01-35.67 10.24-42.15-5.16z"
-    />
-    <path
-     fill="#ffe0de"
-     d="M134.83 156.28c.08.52 1 7.93-1.19 9.6s-7.12-2.27-7.24-4.23-1.79-6 .31-7.06 7.53-1.97 8.12 1.69z"
-    />
-    <path
-     fill="#e0902f"
-     d="M133 151.08l-.25-.16c-2.81-1.65-7.44-1.24-9.77 1.13a11.62 11.62 0 0 0-1.62 2.24 16.83 16.83 0 0 1-1.13 1.73 11.26 11.26 0 0 0 3.15 8 13.24 13.24 0 0 0 6.71 2.05c.92-2.75-.41-5.88.9-8.64a6.3 6.3 0 0 1 2.92-2.82 6.24 6.24 0 0 0-.91-3.53z"
-    />
-    <path
-     fill="#ffe0de"
-     d="M132.13 161.4a3.09 3.09 0 0 0-1.72-1.83c-.73-.3-1.73-.39-2.17.41-1 1.77 1.42 3 2.7 3.5.6.23.86-.74.27-1l-.21-.08h.07a.45.45 0 0 0 .48-.46c.33.12.73-.12.58-.54z"
-    />
-    <path
-     fill="#68e1fd"
-     d="M151.6 158.32l20.63 12.92-3.42 14.55-19.27-13.16 2.06-14.31z"
-    />
-    <path
-     fill="#3129a3"
-     d="M149.54 172.62S139 177.87 139.13 178s18.62 14 18.62 14l11.06-6.18z"
-    />
-    <path
-     fill="#ffe0de"
-     d="M154 183.73s2.8-2.51 3.22-2.71a2.12 2.12 0 0 1 1.82.51c.45.42 1.08.27 1.58.81s1.77 2.44 1.41 2.94-1.37-.71-1.71-1 0 1-.48 1.77-3.39 1.5-4.17 1.35"
-    />
-    <path
-     fill="#fcca4c"
-     d="M140.22 172.62s3.43 11.71 5.44 11.75 8.32-2.09 8.32-2.09 2.72 2.78 2.41 4.54-11.49 8-16.16 6.86-6.93-17.28-6.83-17.59 6.82-3.47 6.82-3.47z"
-    />
-    <path
-     fill="#ffe0de"
-     d="M152.16 176.73a15.26 15.26 0 0 0-1.71-1.12c-.52-.24-1-.51-1.46-.77a1.77 1.77 0 0 0-1.39-.22 36.33 36.33 0 0 0-2.9 2.38c-.56.4-1.13.79-1.72 1.16a.5.5 0 0 0 .39.91c.29-.07.57-.16.85-.26a3.88 3.88 0 0 0 0 .53 2.55 2.55 0 0 0 .32 1.77.48.48 0 0 0 .14.17l-.37.14c-.6.22-.34 1.19.27 1l3.44-1.26a4.86 4.86 0 0 0 2.21-1.26.5.5 0 0 0-.49-.84c-.25.06-.49.14-.73.22.55-.75 1.39-1.65 1.68-1.4a2 2 0 0 0 2.3.08c.54-.54-.42-.96-.83-1.23z"
-    />
-    <path
-     fill="#fcca4c"
-     d="M144.68 177.1s-3.57.47-6.24.54c-1.09 1.48-2.08 3-3 4.6.2 1 .28 2 .5 3a2.13 2.13 0 0 1 0 .58c4.75-.78 9.81-2.74 10.22-3.79.64-1.68-1.48-4.93-1.48-4.93z"
-    />
-    <path
-     fill="#cc781d"
-     d="M125.23 154a2.24 2.24 0 0 0-1.7-1.68c-2.48-.66-5.75.85-6.49 3.41a5.29 5.29 0 0 0 4.07 6.27 5 5 0 0 0 5.44-4.92 4.33 4.33 0 0 0-1.32-3.08z"
-    />
-    <path
-     fill="#3129a3"
-     d="M210.49 246.48c0 .42-.33 5.44.23 6s2 3.42 3.85 3.55 3.51-.24 3.85 1.71-10.32 2.61-10.32 2.61-1.12-4.46-1.6-4.33-.5 4.98-.5 4.98l-2 .42-2-1.32.8-8.56 1.29-7.27"
-    />
-    <path
-     fill="#3129a3"
-     d="M218.48 253.66c-.26 3-.11 7.08 2.74 8.92 1.56 1 4.85.08 5.25 2.31.35 1.95-10.31 2.67-10.31 2.67s-1.15-4.46-1.63-4.32-.43 4.93-.43 4.93l-2 .44-2-1.31.75-8.56 1.25-7.28"
-    />
-    <path
-     fill="#edbab9"
-     d="M201 242.61l-1.57 8.56s2.46 2.53 10.75 2.23 0-15.14 0-15.14"
-    />
-    <ellipse cx="190.26" cy="262.58" fill="#56403d" rx="11.8" ry="4.38" />
-    <path fill="#56403d" d="M188.27 242.96l.46 17.65h3.52v-17.65" />
-    <path
-     fill="#008074"
-     d="M213 217.49q-5.3 1.16-10.69 1.79a.6.6 0 0 0-.41.22q-.62-1.86-1.06-3.79c-.16-.73-1.28-.42-1.12.31a40.46 40.46 0 0 0 2.93 8.45c.32.68 1.32.09 1-.59-.53-1.13-1-2.27-1.43-3.44h.08q5.54-.65 11-1.84c.7-.15.39-1.27-.3-1.11zM203.41 225.68a15.6 15.6 0 0 1-3.58 4.11c-.58.47.25 1.29.82.82a16.5 16.5 0 0 0 3.76-4.34.58.58 0 0 0-1-.59z"
-    />
-    <path
-     fill="#edbab9"
-     d="M196.17 221.9s9.25-7.42 15.39-8.3 9.35 5.79 8.91 10.25-15.14 6.9-15.14 6.9"
-    />
-    <path
-     fill="#ffe0de"
-     d="M201.1 224.37c1.2-1.11 16.9-8.44 22.43-5.09s5.18 5.82 4.48 9.9-6 26-6 26a20 20 0 0 1-12.56.59L211 239.5l-21-11.43"
-    />
-    <path
-     fill="#006d29"
-     d="M189.66 198.14c.83.61 8 1.47 9.5 4.78s3.78 20.34 3.78 20.34-4.3 6.41-11.75 4.13c-2.21-.68-13.67-21.63-14-23.8-.79-5.67 12.47-5.45 12.47-5.45z"
-    />
-    <path
-     fill="#68e1fd"
-     d="M175.52 234.94a18.76 18.76 0 0 1-.89-2.65c-1.76-6.89-2.91-32.76 8.72-31.7 12.76 1.15 8.45 33.1 21.38 32.12s18.23 2.38 12.94 7.39c-10.57 10.01-35.67 10.24-42.15-5.16z"
-    />
-    <path
-     fill="#ffe0de"
-     d="M194.24 188.9c.08.52 1 7.93-1.19 9.6s-7.12-2.27-7.24-4.23-1.79-6 .31-7.06 7.53-1.97 8.12 1.69z"
-    />
-    <path
-     fill="#e0902f"
-     d="M192.15 183.55c-2.81-1.65-7.44-1.24-9.77 1.13a11.62 11.62 0 0 0-1.62 2.24 16.83 16.83 0 0 1-1.13 1.73 11.26 11.26 0 0 0 3.15 8 12.79 12.79 0 0 0 7.93 2c-.24-1-.45-2-.61-3-.26-1.63-.65-3.62.56-5a1.7 1.7 0 0 1 1.73-.58l.08-.05 2.4.15a5.44 5.44 0 0 0-2.72-6.62z"
-    />
-    <path
-     fill="#e0902f"
-     d="M180.71 188c-2.67-4.8-10.76-2.44-11.36 2.64a10 10 0 0 0 1.83 6.85c1.92 2.4 2 7.75-1.32 11.65s-.61 8.59 3.24 10.25 12.19 1.61 11.9-5.13-8.42-15.35-8.3-20 4.8-2 4.8-2"
-    />
-    <path
-     fill="#3129a3"
-     d="M182 188.19a2 2 0 0 0-2-2 2.05 2.05 0 0 0-2 2 18.75 18.75 0 0 0-.06 2.41 2 2 0 0 0 4 0 18.75 18.75 0 0 1 .06-2.41z"
-    />
-    <path
-     fill="#ffe0de"
-     d="M191.54 194a3.09 3.09 0 0 0-1.72-1.83c-.73-.3-1.73-.39-2.17.41-1 1.77 1.42 3 2.7 3.5.6.23.86-.74.27-1l-.21-.08h.07a.45.45 0 0 0 .48-.46c.33.14.73-.1.58-.54z"
-    />
-    <path
-     fill="#68e1fd"
-     d="M211.01 190.94l20.63 12.92-3.42 14.55-19.27-13.16 2.06-14.31z"
-    />
-    <path
-     fill="#3129a3"
-     d="M208.95 205.25s-10.53 5.25-10.41 5.39 18.62 13.95 18.62 13.95l11.06-6.18z"
-    />
-    <path
-     fill="#ffe0de"
-     d="M213.38 216.35s2.8-2.51 3.22-2.71a2.12 2.12 0 0 1 1.82.51c.45.42 1.08.27 1.58.81s1.77 2.44 1.41 2.94-1.37-.71-1.71-1 0 1-.48 1.77-3.39 1.5-4.17 1.35"
-    />
-    <path
-     fill="#006d29"
-     d="M199.63 205.25s3.43 11.75 5.43 11.75 8.32-2.09 8.32-2.09 2.72 2.78 2.41 4.54-11.49 8-16.16 6.86-6.93-17.28-6.83-17.59 6.83-3.47 6.83-3.47z"
-    />
-    <path
-     fill="#ffe0de"
-     d="M211.57 209.35a15.26 15.26 0 0 0-1.71-1.12c-.52-.24-1-.51-1.46-.77a1.77 1.77 0 0 0-1.39-.22 36.33 36.33 0 0 0-2.91 2.42c-.56.4-1.13.79-1.72 1.16a.5.5 0 0 0 .39.91c.29-.07.57-.16.85-.26a3.88 3.88 0 0 0 0 .53 2.55 2.55 0 0 0 .32 1.77.48.48 0 0 0 .14.17l-.37.14c-.6.22-.34 1.19.27 1l3.44-1.26a4.86 4.86 0 0 0 2.21-1.26.5.5 0 0 0-.49-.84c-.25.06-.49.14-.73.22.55-.75 1.39-1.65 1.68-1.4a2 2 0 0 0 2.3.08c.55-.62-.39-1.02-.82-1.27z"
-    />
-    <path
-     fill="#006d29"
-     d="M204.09 209.72s-3.57.47-6.24.54c-1.09 1.48-2.08 3-3 4.6.2 1 .28 2 .5 3a2.13 2.13 0 0 1 0 .58c4.75-.78 9.81-2.74 10.22-3.79.64-1.65-1.48-4.93-1.48-4.93z"
-    />
-    <path
-     fill="#e89c89"
-     d="M250.2 79.78l-5.39-1.05-.08.21a3.84 3.84 0 0 0-1-.41 2.11 2.11 0 0 0-2.56 1.33 1.78 1.78 0 0 0 0 1.26 1.85 1.85 0 0 0-.52 1.34 2.06 2.06 0 0 0 .56 1.42 1.6 1.6 0 0 0 .26 1.91 4.51 4.51 0 0 0 2.27 1c1.42 2.78 4.86 6.45 5.91 5.87 1.37-.76 3.36-4.35 3.36-4.35z"
-    />
-    <path fill="#935a4e" d="M248.8 81.69l-24.58-35.86-.97.42 26.45 42.79" />
-    <path
-     fill="#ffb7a2"
-     d="M254.94 87.59s-.94-7.59-2.36-8.45-6.07-2.78-6.94-1.77-1.67 3-.83 3.79 2.46-1.16 2.46-1.16l1.41 1s-1.84 1.59-1.73 5.38a9.76 9.76 0 0 0 2.75 6.29z"
-    />
-    <path
-     fill="#27d9d3"
-     d="M274.19 123.12l-2.37-35.53c-1 .87-4.3 21-5.94 34.7 2.78.15 5.54.48 8.31.83z"
-    />
-    <path
-     fill="#ffb7a2"
-     d="M328.4 87.18l.6-1.08a1 1 0 0 1-.41-1.24c.51-1.17 1.05-2.32 1.59-3.48a.94.94 0 0 1 0-.43 26 26 0 0 0 .88-8c-.23-2.16-.88-4.23-1.09-6.39-.08-.77.81-1 1.38-.79a2.23 2.23 0 0 1 1.09 1.15 7.47 7.47 0 0 1 1.25 3.26 5 5 0 0 1 .37-1.46q.22-.88.43-1.76a12.87 12.87 0 0 1 .92-2.7c.41-.81 1.33-1.18 1.79-.65.7.8 0 2.25-.11 3.18s-.28 1.75-.52 2.69c.36-.87 1.1-2.39 1.58-3.44.37-.81 1.42-3.45 2.58-3.09s.61 2.16.29 3l-.91 2.45c-.27.73-.57 1.68-.88 2.35 0 0 1.22-1.78 1.83-2.73.49-.75 1-1.84 2.11-1.51.71.22.66 1.27.33 1.9-.88 1.67-1.86 3.27-2.88 4.84.66-.61 1.29-1.24 2-1.83a1 1 0 0 1 1.56.8.79.79 0 0 1-.14.49l-8.44 12.13a17.75 17.75 0 0 1-3.43 5.22c-.81.85-2.11-.44-1.29-1.29l.11-.13a.91.91 0 0 1-.08-.14 1 1 0 0 1-.65-.49h-.06a.92.92 0 0 1-.29.13.91.91 0 0 1-1.51-.96z"
-    />
-    <path
-     fill="#42342d"
-     d="M334.9 70.92a5.53 5.53 0 0 1 2.07 4.8c0 .41-.68.42-.65 0a4.94 4.94 0 0 0-1.88-4.35c-.33-.25.13-.7.46-.45zM335.79 69.93q.7-2.55 1.17-5.16c.07-.41.7-.24.62.17q-.47 2.6-1.17 5.16a.32.32 0 0 1-.62-.17zM338.52 71.36a37.08 37.08 0 0 0 2.13-5.87.32.32 0 0 1 .62.17 38.1 38.1 0 0 1-2.19 6c-.18.41-.73.08-.56-.3zM339.53 73.9a24.35 24.35 0 0 0 3.47-5.24c.18-.37.74 0 .56.33a25 25 0 0 1-3.56 5.37c-.29.32-.75-.14-.47-.46z"
-    />
-    <path
-     fill="#159e97"
-     d="M310.32 150.91A129.2 129.2 0 0 0 301 128l-25.52-.24c-.56 3.48-1.56 6.88-2.18 10.41a51.36 51.36 0 0 0-.74 11.37c7.21 4.06 16.68 9.57 23.53 13.51l19 10.9c-1.61-7.56-2.66-15.57-4.77-23.04z"
-    />
-    <path
-     fill="#27d9d3"
-     d="M302.32 105.15q-.56-8.4-1.57-16.78c-3.19-3.95-7.57-7-12.38-7.64s-11.82.62-15.1 4.49c-3.1 3.69-2.58 9-2.39 13.43.11 2.7.25 5.4.4 8.1 1.41 7.25 3.45 14.36 3.88 21.76 7.09 3.32 17.79 1.62 25.8-.54.47-7.61 1.04-15.21 1.36-22.82z"
-    />
-    <path
-     fill="#27d9d3"
-     d="M289 81.47a22.78 22.78 0 0 1 10.81 3.36c2.48 1.62 3.9 4.24 6.58 5.58 3.6 1.81 8.88 1.79 11.84-1.26a65.83 65.83 0 0 0 4.5-5.49c1.11-1.42 5.67-8.82 8.07-7.91 2.81 1.06 6.86 5 7 7.25s-11.72 25.68-22.23 27.33a28 28 0 0 1-18.78-4.16"
-    />
-    <path fill="#fcaa97" d="M291.58 77.36v6.71a13.67 13.67 0 0 1-8.1 0V78" />
-    <path
-     fill="#27d9d3"
-     d="M271.33 108s-14.61-20.89-16.39-20.39-7.42 3.29-7 5.55 11.94 35.48 27.65 33.55"
-    />
-    <path
-     fill="#159e97"
-     d="M291.73 80.2a11.25 11.25 0 0 1-2.23 3.44c-1 1.13-2.44.76-4 .85l1.06 8.19 7.09-10zM285.46 84.49l-3.46 7.6-.54-7.09 2.31-4.6s-.77 3.88 1.69 4.09z"
-    />
-    <path
-     fill="#ffb7a2"
-     d="M277.56 62c-.47.12-3.44 20.1 4.85 20.26s9.89-6.36 10.14-8-2.39-6.74-3.2-8.64-9.76-4.17-11.79-3.62z"
-    />
-    <path
-     fill="#edb033"
-     d="M276.53 59.93c2.5-4.51 12.18-5.54 18.91.64a9.89 9.89 0 0 1 2.72 6.78 14.21 14.21 0 0 1-8 11.81v-.08c0-.08 0-.25.07-.37s0-.18.05-.27c.71-3.49.55-7.89-1.1-11s-5-4-8.25-4.55a10.33 10.33 0 0 0-3.23-.32l-1.07 5.13a6.92 6.92 0 0 1-.1-7.77z"
-    />
-    <circle
-     cx="301.32"
-     cy="62.39"
-     r="6.38"
-     fill="#edb033"
-     transform="rotate(-86.02 301.336 62.38)"
-    />
-    <path
-     fill="#edb033"
-     d="M276.59 66.08h7.86l.73-2.36.68 2.93 2.24.14s.11 12.07 1.83 12.34 4.13-4.92 4.39-8-1.41-9.57-7.08-11.33c-7.14-2.16-10.32.69-10.65 6.28z"
-    />
-    <path
-     fill="#159e97"
-     d="M272 95.94c.06-.68-1-.67-1.05 0-1.08 11.71 4.05 22.75 4.42 34.36 0 .68 1.08.68 1.05 0-.42-11.61-5.53-22.66-4.42-34.36zM304.38 108.58a12.67 12.67 0 0 1-6.23-3.12c-.5-.46-1.25.29-.75.75a13.7 13.7 0 0 0 6.7 3.39c.66.13.9-.89.28-1.02z"
-    />
-    <path
-     fill="#00a99d"
-     d="M184.72 47.37l35.42 18.94c1.7.91 3.22-1.68 1.51-2.59l-35.41-18.94c-1.7-.91-3.22 1.68-1.51 2.59zM184.72 59.36l35.42 18.94c1.7.91 3.22-1.68 1.51-2.59l-35.41-18.94c-1.7-.91-3.22 1.68-1.51 2.59z"
-    />
-    <ellipse
-     cx="192.74"
-     cy="84.3"
-     fill="#68e1fd"
-     rx="7.16"
-     ry="11.26"
-     transform="rotate(-18.56 192.7 84.299)"
-    />
-    <ellipse
-     cx="212.62"
-     cy="95.44"
-     fill="#68e1fd"
-     rx="7.16"
-     ry="11.26"
-     transform="rotate(-18.56 212.558 95.43)"
-    />
-   </svg>
-  </>
- );
-}
