--- src/components/svg/I2dIllustrations.jsx
@@ -1,23 +1,0 @@
-export default function I2dIllustrations() {
- return (
-  <svg
-   width="350"
-   height="411"
-   viewBox="0 0 600 601"
-   fill="none"
-   xmlns="http://www.w3.org/2000/svg"
-  >
-   <path
-    d="M172.852 593.124C421.869 598.115 509.248 462.116 594.449 1C428.004 367.284 345.249 564.342 172.852 593.124Z"
-    fill="#48AF48"
-    stroke="#48AF48"
-   />
-   <path
-    fillRule="evenodd"
-    clipRule="evenodd"
-    d="M480.898 4.80557C358.642 15.3167 268.038 36.9667 192.035 73.8278C139.088 99.5056 134.125 103.278 130.887 120.278C123.414 159.506 139.719 176.667 151.112 141.567C158.41 119.083 171.534 100.611 166.12 120.445C151.608 173.617 159.16 183.772 181.612 141.278C197.912 110.433 198.808 111.206 187.873 146.695C178.527 177.039 186.558 189.472 201.335 167.528C206.202 160.295 214.685 149.245 220.189 142.967C228.13 133.911 227.622 136.983 217.719 157.822C201.42 192.122 208.706 198.133 232.868 170.317C255.388 144.389 258.518 148.4 238.457 177.478C218.514 206.389 229.117 213.517 258.021 190.639C284.461 169.706 288.426 174.945 263.763 198.222C254.457 207.006 246.843 216.222 246.843 218.706C246.843 226.55 260.424 223.378 275.043 212.111C297.912 194.489 301.347 199.445 280.767 220.361C259.708 241.761 261.591 250.961 285.127 241.711C319.485 228.211 323.765 227.583 309.186 238.189C301.601 243.706 290.18 252.784 283.807 258.35C272.307 268.406 272.352 268.461 289.142 265.428C298.448 263.75 312.407 260.861 320.162 259.006C330.466 256.539 328.339 259.117 312.249 268.595C300.14 275.722 290.738 284.056 291.347 287.111C291.957 290.167 304.404 292.145 319 291.506L345.541 290.345L328.621 298.2C319.316 302.517 308.532 306.167 304.652 306.306C300.777 306.445 297.602 308.922 297.602 311.822C297.602 314.717 310.929 317.934 327.211 318.972L356.821 320.856L333.63 325.406C287.76 334.4 285.708 350.734 330.364 351.389C350.335 351.684 354.057 352.95 345.541 356.556C339.337 359.184 326.01 361.506 315.932 361.722C288.928 362.295 293.643 371.956 323.049 376.3C337.047 378.367 350.064 382.55 351.97 385.595C354.226 389.184 345.112 389.728 325.993 387.139C299.649 383.567 296.745 384.189 298.482 393C299.649 398.917 310.139 406.05 324.741 410.839C355.631 420.978 352.405 424.667 318.171 418.367C282.476 411.8 279.041 420.673 311.753 434.928C341.982 448.1 342.044 455.306 311.848 445.495C300.653 441.856 287.394 440.428 282.38 442.323C275.251 445.017 278.376 449.473 296.717 462.756C309.609 472.095 320.162 480.739 320.162 481.967C320.162 483.189 308.797 478.75 294.906 472.1C262.178 456.439 255.777 463.273 282.696 485.128C307.167 504.995 306.282 513.4 281.545 496.05C265.709 484.939 246.843 480.4 246.843 487.7C246.843 489.55 254.671 500.8 264.236 512.7C285.268 538.856 282.295 540.267 255.309 516.928C229.399 494.517 224.081 500.089 243.623 529.173C263.475 558.728 258.839 563.173 236.285 536.217C223.618 521.084 217.082 516.984 212.62 521.378C208.154 525.778 209.727 533.778 218.3 550.339C232.732 578.2 226.314 580.462 208.723 553.712C201.713 543.05 193.467 534.334 190.394 534.334C183.259 534.334 183.237 534.112 193.326 562.278C204.916 594.65 199.412 596.934 184.19 566.067C170.524 538.362 154.107 535.134 159.989 561.3C168.353 598.512 166.65 597.117 199.214 593.35C234.7 589.239 302.954 565.939 331.503 548.184C388.178 512.945 432.451 434.734 465.016 312.328C479.561 257.661 479.742 257.345 489.967 269.406C496.064 276.606 500.339 278.595 500.452 274.284C500.553 270.284 502.809 268.334 505.465 269.95C508.116 271.567 509.092 266.711 507.626 259.167C505.967 250.622 507.563 245.445 511.861 245.445C515.922 245.445 517.315 241.717 515.245 236.406C513.305 231.439 514.461 224.122 517.805 220.15C521.15 216.183 523.327 209.622 522.65 205.578C521.973 201.533 522.881 194.472 524.675 189.889C526.468 185.306 534.522 157.806 542.576 128.778C550.63 99.7501 589.764 26.6831 600 6.06089L594.457 1L548.954 1.93333C523.93 2.45 493.306 3.7389 480.898 4.80557ZM112.518 118.872C104.769 126.506 103.449 156.556 110.86 156.556C115.885 156.556 125.383 118.389 121.345 114.411C120.16 113.25 116.19 115.256 112.518 118.872ZM84.6966 138.333C76.2312 141.311 75.1992 156.556 83.4616 156.556C89.2819 156.556 97.8263 140.989 93.8558 137.633C92.694 136.656 88.5769 136.972 84.6966 138.333ZM51.224 184.372C44.4505 187.061 36.1373 194.35 32.759 200.572C26.1152 212.806 30.1309 237.772 39.6792 243.589C43.441 245.878 44.3829 242.65 42.2679 234.689C39.0757 222.672 39.3633 222.567 49.8422 231.906C60.3042 241.234 60.7271 241.1 60.7271 228.578C60.7271 215.922 60.9697 215.856 69.0065 226.328C77.0095 236.75 77.3084 236.645 77.8272 223.222L78.363 209.333L84.409 223.222L90.4549 237.111L94.7695 223.222C97.595 214.117 99.2757 212.683 99.6423 219.056C99.9525 224.406 102.558 228.778 105.434 228.778C108.305 228.778 111.819 224.406 113.24 219.056C115.18 211.739 115.981 212.428 116.472 221.833C117.256 236.861 121.598 237.611 132.472 224.611C140.221 215.35 140.458 215.35 137.503 224.611C135.794 229.961 136.86 234.333 139.86 234.333C142.866 234.333 145.325 231.833 145.325 228.778C145.325 225.722 147.626 223.222 150.435 223.222C153.249 223.222 154.219 228.222 152.601 234.333C148.963 248.039 155.505 248.711 166.148 235.722C173.401 226.867 173.762 227.233 170.186 239.845C166.554 252.645 167.022 253.067 176.345 245.445C185.617 237.867 186.107 238.206 182.419 249.656C178.364 262.228 186.513 267.317 193.321 256.467C195.272 253.361 199.023 252.133 201.656 253.739C207.731 257.433 224.284 241.456 224.284 231.895C224.284 222.75 184.996 196.978 155.347 186.672C130.38 177.995 70.6195 176.672 51.224 184.372ZM62.1372 251.172C55.155 253.939 49.4475 258.639 49.4475 261.622C49.4475 274.578 78.2333 298.178 104.199 306.517L132.393 315.572L154.056 297.656C173.57 281.522 174.912 278.784 167.575 270.067C149.442 248.539 93.9235 238.589 62.1372 251.172ZM224.284 273.222C216.14 279.5 213.455 284.139 217.922 284.211C224.47 284.317 224.475 285.745 217.967 293.467C211.092 301.634 211.769 302.195 224.329 298.734C237.131 295.211 237.504 295.567 228.514 302.711C215.299 313.217 216.083 317.539 231.333 318.311C240.882 318.795 241.581 319.584 234.153 321.495C228.722 322.895 224.284 326.172 224.284 328.778C224.284 331.384 228.722 334.661 234.153 336.061C243.555 338.478 243.555 338.634 234.153 339.245C219.619 340.195 222.124 350.939 236.973 351.35C249.313 351.695 249.274 351.867 235.563 357.489C222.129 363.006 221.999 363.495 232.743 367.861C242.348 371.761 242.557 372.5 234.153 372.834C228.722 373.045 224.284 374.822 224.284 376.778C224.284 378.734 228.722 383.917 234.153 388.3C243.307 395.689 243.098 396.072 231.333 393.6C216.106 390.395 215.147 394.217 228.514 404.845C237.611 412.073 237.391 412.417 225.694 409.245C211.684 405.45 209.163 411.473 220.054 422.734C225.818 428.7 224.532 429.272 213.004 425.884L198.904 421.745L210.184 434.406L221.464 447.061L208.289 440.245C192.565 432.111 186.017 436.278 195.971 448.089C199.897 452.745 201.132 456.556 198.718 456.556C196.31 456.556 198.932 461.556 204.544 467.667C218.751 483.128 227.149 481.617 243.346 460.7C280.508 412.706 288.104 330.511 260.272 277.5C250.757 259.372 243.397 258.495 224.284 273.222ZM167.8 316.95C152.934 329.178 148.173 337.467 147.824 351.722C147.576 362.022 145.574 374.284 143.385 378.978C140.842 384.434 146.064 396.056 157.874 411.206L176.345 434.9L188.848 419.478C206.636 397.539 213.342 361.289 205.384 330.067C201.82 316.078 196.299 303.795 193.112 302.767C189.926 301.739 178.539 308.122 167.8 316.95ZM90.2744 332.478C67.1002 341.672 72.0577 378.778 96.4559 378.778C112.738 378.778 122.766 368.9 122.766 352.861C122.766 342.767 107.256 327.222 99.1686 329.217C98.1872 329.456 94.1829 330.928 90.2744 332.478ZM49.6336 410.034C40.2263 417.828 32.5277 428.923 32.5277 434.684C32.5277 446.6 54.3315 459.384 83.9916 464.867C104.154 468.589 147.987 460.239 153.379 451.645C154.969 449.111 148.247 436.75 138.439 424.178C124.875 406.8 117.769 402.022 108.773 404.25C102.271 405.856 90.1504 404.628 81.8429 401.517C69.0517 396.723 64.1112 398.034 49.6336 410.034ZM13.0589 473.923C4.89802 483.473 -0.922358 494.1 0.12102 497.534C1.15876 500.967 14.593 510.017 29.9616 517.645C65.2221 535.134 123.189 536.756 163.655 521.373C193.806 509.917 197.568 499.1 178.736 478.006C172.295 470.789 168.861 468.639 171.105 473.223C174.41 479.984 173.75 480.245 167.586 474.611C156.227 464.228 151.704 466.184 154.699 480.167C157.062 491.178 156.628 491.511 151.072 482.945C142.195 469.261 134.587 470.856 137.813 485.723C140.452 497.911 140.3 497.95 131.643 487.439C122.997 476.934 122.754 477.006 122.41 490.217L122.05 503.778L116.376 489.889L110.697 476L102.829 489.889L94.9669 503.778L94.4085 489.889L93.8502 476L87.8042 489.889L81.7583 503.778L78.0021 489.889C74.3644 476.456 74.026 476.361 67.6529 487.111C61.3475 497.756 61.0655 497.65 60.8964 484.661C60.7328 471.456 60.4959 471.384 51.8499 481.884C43.1927 492.395 43.0405 492.356 45.68 480.167C48.7086 466.2 44.129 464.189 33.0241 474.611C26.9838 480.278 26.324 479.256 29.4541 469.056C34.6428 452.139 30.6497 453.323 13.0589 473.923ZM4.32841 542.339C4.32841 546.745 6.86632 551.889 9.96826 553.778C13.0702 555.667 15.6081 552.062 15.6081 545.773C15.6081 539.484 13.0702 534.334 9.96826 534.334C6.86632 534.334 4.32841 537.934 4.32841 542.339ZM24.0849 556.506C20.2385 568.434 24.3162 578.778 32.8606 578.778C35.782 578.778 38.1677 571.278 38.1677 562.112C38.1677 543.25 29.471 539.789 24.0849 556.506ZM49.267 565.356C43.0575 581.45 46.1425 589.889 58.24 589.889C63.7671 589.889 66.3671 583.673 66.3671 570.445C66.3671 547.617 57.1685 544.878 49.267 565.356ZM134.125 563.5C134.271 587.35 140.78 601 152.003 601C160.824 601 162.222 598.523 159.053 588.5C150.204 560.528 145.32 551 139.826 551C136.646 551 134.079 556.623 134.125 563.5ZM81.4087 560.262C73.7836 567.767 77.4832 595.445 86.1066 595.445C92.0172 595.445 94.5665 589.578 94.5665 576C94.5665 557.617 89.8346 551.956 81.4087 560.262ZM105.846 578.778C105.846 597.295 107.724 601 117.126 601C128.891 601 131.079 591.945 124.853 569.056C118.948 547.35 105.846 554.056 105.846 578.778Z"
-    fill="#12633A"
-   />
-  </svg>
- );
-}
