--- src/components/svg/MathsIllustration.jsx
@@ -1,374 +1,0 @@
-export default function MathsIllustration() {
- return (
-  <svg
-   width="600"
-   viewBox="0 0 1051 720"
-   fill="none"
-   xmlns="http://www.w3.org/2000/svg"
-  >
-   <path
-    opacity="0.2"
-    d="M127.458 325.302C60.4585 194.887 171.423 91.9012 229.749 68.2831C281.984 51.2691 331.655 63.6488 370.562 84.1094C414.886 112.11 446.32 176.554 454.09 248.617C460.306 306.268 484.837 359.014 515.657 386.138C532.995 401.397 597.261 447.786 605.358 461.903C705.362 575.088 581.181 671.812 536.207 676.265C502.631 679.59 408.678 643.555 385.797 547.676C366.515 466.881 267.698 441.35 247.431 431.521C209.836 413.289 163.439 395.339 127.458 325.302Z"
-    fill="#FF85B0"
-   />
-   <path
-    opacity="0.2"
-    d="M735.245 124.606C858.008 52.9994 969.843 133.823 989.322 210.304C1008.5 259.982 999.086 308.702 981.297 347.421C956.49 391.774 895.741 425.539 826.391 436.792C770.911 445.795 721.127 472.272 696.469 503.509L419.45 506.858C429.48 465.729 463.521 406.856 533.383 386.181C610.589 363.332 630.183 266.361 638.647 246.234C654.347 208.9 669.317 163.062 735.245 124.606Z"
-    fill="#0551D8"
-   />
-   <ellipse
-    opacity="0.2"
-    cx="516.59"
-    cy="669.263"
-    rx="275.742"
-    ry="50.1349"
-    fill="#FED601"
-   />
-   <circle
-    cx="777.813"
-    cy="213.973"
-    r="41.8719"
-    transform="rotate(-104.464 777.813 213.973)"
-    fill="white"
-   />
-   <mask id="path-5-inside-1_725_3488" fill="white">
-    <path d="M763.01 191.492C755.143 195.197 748.552 201.151 744.069 208.603C739.586 216.054 737.414 224.667 737.826 233.352C738.238 242.038 741.217 250.407 746.386 257.399C751.554 264.392 758.68 269.696 766.862 272.638C775.045 275.581 783.916 276.032 792.355 273.933C800.793 271.834 808.42 267.28 814.27 260.847C820.12 254.413 823.931 246.389 825.221 237.79C826.511 229.191 825.222 220.402 821.517 212.535L781.742 231.267L763.01 191.492Z" />
-   </mask>
-   <path
-    d="M763.01 191.492C755.143 195.197 748.552 201.151 744.069 208.603C739.586 216.054 737.414 224.667 737.826 233.352C738.238 242.038 741.217 250.407 746.386 257.399C751.554 264.392 758.68 269.696 766.862 272.638C775.045 275.581 783.916 276.032 792.355 273.933C800.793 271.834 808.42 267.28 814.27 260.847C820.12 254.413 823.931 246.389 825.221 237.79C826.511 229.191 825.222 220.402 821.517 212.535L781.742 231.267L763.01 191.492Z"
-    stroke="black"
-    strokeWidth="8.37439"
-    mask="url(#path-5-inside-1_725_3488)"
-   />
-   <mask id="path-6-inside-2_725_3488" fill="white">
-    <path d="M827.531 195.819C825.071 190.595 821.606 185.908 817.334 182.023C813.063 178.139 808.068 175.134 802.635 173.18C797.202 171.226 791.437 170.361 785.67 170.635C779.903 170.909 774.246 172.316 769.023 174.776L787.755 214.551L827.531 195.819Z" />
-   </mask>
-   <path
-    d="M827.531 195.819C825.071 190.595 821.606 185.908 817.334 182.023C813.063 178.139 808.068 175.134 802.635 173.18C797.202 171.226 791.437 170.361 785.67 170.635C779.903 170.909 774.246 172.316 769.023 174.776L787.755 214.551L827.531 195.819Z"
-    stroke="black"
-    strokeWidth="8.37439"
-    mask="url(#path-6-inside-2_725_3488)"
-   />
-   <rect
-    x="340.196"
-    y="154.294"
-    width="74.6067"
-    height="73.0841"
-    transform="rotate(15.3186 340.196 154.294)"
-    fill="white"
-   />
-   <path
-    d="M359.471 143.411C358.92 142.579 357.799 142.352 356.967 142.903L343.414 151.888C342.582 152.439 342.355 153.56 342.906 154.392C343.457 155.224 344.578 155.451 345.41 154.899L357.458 146.913L365.444 158.961C365.995 159.792 367.116 160.02 367.948 159.468C368.78 158.917 369.007 157.796 368.456 156.964L359.471 143.411ZM338.136 233.135C337.937 234.113 338.569 235.067 339.547 235.265C340.525 235.463 341.479 234.831 341.677 233.853L338.136 233.135ZM356.194 144.05L338.136 233.135L341.677 233.853L359.736 144.768L356.194 144.05Z"
-    fill="black"
-   />
-   <path
-    d="M416.921 233.438C417.751 232.883 417.973 231.761 417.418 230.932L408.373 217.419C407.817 216.589 406.695 216.367 405.866 216.922C405.037 217.477 404.815 218.599 405.37 219.429L413.41 231.44L401.398 239.48C400.569 240.036 400.347 241.158 400.902 241.987C401.457 242.816 402.579 243.038 403.408 242.483L416.921 233.438ZM328.098 212.701C327.119 212.508 326.168 213.144 325.975 214.123C325.781 215.102 326.417 216.052 327.396 216.246L328.098 212.701ZM416.267 230.164L328.098 212.701L327.396 216.246L415.565 233.709L416.267 230.164Z"
-    fill="black"
-   />
-   <line
-    x1="355.276"
-    y1="220.707"
-    x2="354.833"
-    y2="222.948"
-    stroke="black"
-    strokeWidth="2.28388"
-    strokeLinecap="round"
-   />
-   <line
-    x1="343.151"
-    y1="203.561"
-    x2="345.391"
-    y2="204.005"
-    stroke="black"
-    strokeWidth="2.28388"
-    strokeLinecap="round"
-   />
-   <line
-    x1="367.974"
-    y1="223.221"
-    x2="367.53"
-    y2="225.462"
-    stroke="black"
-    strokeWidth="2.28388"
-    strokeLinecap="round"
-   />
-   <line
-    x1="345.663"
-    y1="190.866"
-    x2="347.904"
-    y2="191.309"
-    stroke="black"
-    strokeWidth="2.28388"
-    strokeLinecap="round"
-   />
-   <line
-    x1="381.416"
-    y1="225.883"
-    x2="380.972"
-    y2="228.123"
-    stroke="black"
-    strokeWidth="2.28388"
-    strokeLinecap="round"
-   />
-   <line
-    x1="348.326"
-    y1="177.423"
-    x2="350.566"
-    y2="177.867"
-    stroke="black"
-    strokeWidth="2.28388"
-    strokeLinecap="round"
-   />
-   <line
-    x1="394.111"
-    y1="228.396"
-    x2="393.667"
-    y2="230.636"
-    stroke="black"
-    strokeWidth="2.28388"
-    strokeLinecap="round"
-   />
-   <line
-    x1="350.839"
-    y1="164.728"
-    x2="353.08"
-    y2="165.171"
-    stroke="black"
-    strokeWidth="2.28388"
-    strokeLinecap="round"
-   />
-   <path
-    d="M357.965 207.559C378.857 212.548 414.091 211.924 420.951 162.969"
-    stroke="black"
-    strokeWidth="3.11812"
-    strokeLinecap="round"
-   />
-   <path
-    d="M181.898 256.914C188.752 224.883 164.128 187.424 150.96 172.698L88.0429 254.976L181.898 256.914Z"
-    fill="white"
-   />
-   <path
-    d="M165.824 173.548L89.2729 243.913L191.398 248.213"
-    stroke="black"
-    strokeWidth="3.85061"
-    strokeLinecap="round"
-    strokeLinejoin="round"
-   />
-   <path
-    d="M141.779 225.894C142.114 225.928 142.371 226.027 142.552 226.193C142.732 226.359 142.816 226.594 142.803 226.9C142.8 226.989 142.78 227.101 142.745 227.238C142.709 227.375 142.669 227.511 142.624 227.648C142.579 227.784 142.534 227.911 142.489 228.027C142.445 228.134 142.408 228.207 142.376 228.245C142.29 228.419 142.177 228.523 142.037 228.556C141.908 228.59 141.778 228.62 141.649 228.644C141.52 228.658 141.409 228.698 141.318 228.763C141.226 228.819 141.176 228.955 141.167 229.172C141.161 229.32 141.178 229.503 141.218 229.722C141.258 229.941 141.303 230.175 141.351 230.424C141.4 230.673 141.444 230.927 141.482 231.185C141.521 231.443 141.536 231.676 141.527 231.883C141.503 232.455 141.196 232.817 140.607 232.97C140.397 233.031 140.222 233.078 140.083 233.111C139.943 233.145 139.823 233.169 139.724 233.185C139.625 233.201 139.54 233.207 139.471 233.204C139.402 233.211 139.333 233.213 139.264 233.21C139.106 233.203 138.971 233.133 138.858 233C138.745 232.867 138.649 232.685 138.57 232.455C138.49 232.224 138.418 231.945 138.353 231.616C138.298 231.288 138.244 230.925 138.191 230.528C138.121 230.08 138.058 229.712 138.001 229.424C137.954 229.135 137.889 228.91 137.807 228.749C137.735 228.578 137.646 228.46 137.54 228.397C137.445 228.324 137.318 228.274 137.162 228.247L134.361 227.907C134.076 227.876 133.752 227.837 133.388 227.792C133.035 227.738 132.667 227.673 132.285 227.598C131.913 227.513 131.548 227.409 131.187 227.285C130.837 227.152 130.523 226.99 130.245 226.801C129.976 226.612 129.758 226.39 129.591 226.136C129.434 225.873 129.363 225.564 129.378 225.209C129.389 224.952 129.439 224.707 129.528 224.474C129.617 224.231 129.726 223.999 129.854 223.777C129.982 223.545 130.12 223.319 130.267 223.098C130.425 222.877 130.572 222.661 130.709 222.45C131.259 221.604 131.861 220.795 132.515 220.022C133.18 219.24 133.826 218.453 134.452 217.659C134.712 217.354 134.849 217.019 134.865 216.654C134.871 216.497 134.853 216.333 134.811 216.163C134.769 215.984 134.752 215.82 134.758 215.672C134.766 215.465 134.828 215.305 134.941 215.191C135.064 215.078 135.211 214.995 135.381 214.943C135.561 214.891 135.755 214.86 135.963 214.849C136.171 214.838 136.373 214.836 136.57 214.845C137.083 214.866 137.453 214.986 137.681 215.203C137.919 215.41 138.077 215.649 138.155 215.919C138.242 216.189 138.28 216.458 138.269 216.724C138.268 216.98 138.285 217.174 138.319 217.304C138.648 218.631 138.983 219.944 139.323 221.242C139.664 222.531 139.979 223.838 140.269 225.163C140.319 225.393 140.411 225.555 140.545 225.649C140.679 225.744 140.859 225.806 141.085 225.835L141.779 225.894ZM135.686 225.296C135.793 225.321 135.911 225.34 136.039 225.356C136.167 225.371 136.28 225.381 136.378 225.385C136.556 225.392 136.696 225.359 136.797 225.284C136.91 225.2 136.969 225.079 136.976 224.921C136.978 224.872 136.975 224.822 136.967 224.773C136.96 224.713 136.947 224.653 136.93 224.593C136.87 224.373 136.811 224.149 136.751 223.919C136.702 223.68 136.643 223.431 136.575 223.171C136.517 222.912 136.445 222.637 136.358 222.347C136.282 222.047 136.182 221.722 136.059 221.371C136.033 221.281 135.978 221.17 135.895 221.039C135.811 220.907 135.691 220.837 135.533 220.831C135.356 220.823 135.178 220.929 135.001 221.149C134.636 221.618 134.281 222.077 133.936 222.526C133.601 222.977 133.28 223.437 132.974 223.908C132.867 224.101 132.811 224.262 132.805 224.39C132.8 224.528 132.839 224.643 132.924 224.736C133.019 224.829 133.145 224.893 133.301 224.929L135.686 225.296ZM160.128 216.652C160.472 216.696 160.748 216.811 160.958 216.998C161.168 217.174 161.31 217.442 161.383 217.801L161.457 218.159C161.474 218.239 161.485 218.319 161.492 218.398C161.498 218.467 161.5 218.541 161.497 218.62C161.487 218.867 161.399 219.061 161.235 219.202C161.081 219.344 160.866 219.418 160.589 219.426C159.792 219.363 159.073 219.308 158.433 219.262C157.793 219.205 157.168 219.164 156.557 219.138C155.817 219.107 155.092 219.091 154.381 219.091C153.67 219.091 152.967 219.115 152.274 219.165C151.918 219.18 151.734 219.325 151.722 219.601C151.714 219.789 151.866 219.948 152.176 220.08C152.865 220.365 153.579 220.662 154.317 220.97C155.064 221.278 155.796 221.61 156.512 221.966C157.228 222.322 157.908 222.706 158.552 223.118C159.197 223.521 159.756 223.964 160.229 224.448C160.713 224.933 161.085 225.462 161.348 226.036C161.62 226.611 161.742 227.238 161.713 227.918C161.681 228.677 161.494 229.361 161.152 229.969C160.811 230.577 160.34 231.095 159.739 231.524C159.148 231.944 158.434 232.259 157.595 232.471C156.767 232.683 155.844 232.767 154.829 232.725C153.942 232.687 153.064 232.547 152.195 232.303C151.326 232.059 150.551 231.715 149.868 231.271C149.186 230.818 148.636 230.261 148.22 229.602C147.813 228.943 147.628 228.184 147.664 227.327C147.679 226.952 147.728 226.623 147.809 226.34C147.89 226.047 148.006 225.766 148.156 225.496C148.307 225.186 148.494 224.972 148.716 224.852C148.948 224.734 149.197 224.68 149.464 224.691L150.202 224.752C150.901 224.811 151.24 225.067 151.221 225.521C151.219 225.57 151.212 225.624 151.2 225.683C151.187 225.742 151.17 225.8 151.148 225.859C151.014 226.228 150.94 226.571 150.927 226.886C150.904 227.419 151.023 227.883 151.283 228.279C151.553 228.666 151.905 228.992 152.338 229.257C152.772 229.512 153.262 229.706 153.81 229.837C154.357 229.969 154.897 230.046 155.43 230.068C156.465 230.112 157.265 229.983 157.831 229.681C158.397 229.369 158.693 228.917 158.718 228.325C158.742 227.754 158.557 227.232 158.162 226.761C157.767 226.29 157.189 225.822 156.428 225.355C155.668 224.888 154.735 224.405 153.63 223.904C152.525 223.403 151.275 222.832 149.88 222.19C148.844 221.702 148.346 220.965 148.388 219.979C148.391 219.91 148.4 219.802 148.416 219.655C148.432 219.507 148.484 219.233 148.57 218.832C148.684 218.214 148.917 217.74 149.267 217.409C149.627 217.069 150.119 216.857 150.745 216.775C152.125 216.606 153.384 216.501 154.521 216.46C155.669 216.419 156.73 216.42 157.706 216.461C158.564 216.497 159.371 216.56 160.128 216.652Z"
-    fill="black"
-   />
-   <circle
-    cx="170.037"
-    cy="215.605"
-    r="2.39101"
-    transform="rotate(2.41112 170.037 215.605)"
-    stroke="black"
-    strokeWidth="2.39101"
-   />
-   <path
-    d="M334.543 515.899L697.846 515.899L680.806 687.404L356.697 678.312L334.543 515.899Z"
-    fill="#FED601"
-   />
-   <path
-    d="M671.159 487.17H362.826L342.643 376.45L683.271 380.187L671.159 487.17Z"
-    fill="#FF85B0"
-   />
-   <path
-    d="M691.372 508.478H341.81L361.31 488.413H668.021L691.372 508.478Z"
-    fill="#0551D8"
-   />
-   <path
-    d="M334.302 508.478L698.45 508.478L688.343 678.312H349.739L334.302 508.478Z"
-    stroke="black"
-    strokeWidth="4.18077"
-    strokeLinejoin="round"
-   />
-   <path
-    d="M698.449 508.478H334.302L354.939 488.413H677.812L698.449 508.478Z"
-    stroke="black"
-    strokeWidth="4.18077"
-    strokeLinejoin="round"
-   />
-   <path
-    d="M677.956 488.413H354.95L342.183 382.853H685.353L677.956 488.413Z"
-    stroke="black"
-    strokeWidth="4.18077"
-    strokeLinejoin="round"
-   />
-   <path
-    d="M295.931 441.742C297.743 442.21 299.482 442.111 300.92 440.783C302.152 439.639 302.961 438.261 302.756 436.477C302.659 436.33 302.587 436.158 302.466 436.084C300.714 434.891 299.893 434.977 298.721 436.859C298.552 437.129 298.323 437.351 298.154 437.56C297.006 437.363 296.535 436.748 296.644 435.653C296.801 434.03 297.393 432.726 298.782 431.742C300.775 430.327 302.949 429.687 305.341 429.687C306.609 429.687 307.576 430.204 308.566 431.053C310.837 433.009 311.586 437.917 310.547 440.598C308.723 445.285 305.619 448.557 301.222 450.697C300.207 451.189 299.217 451.755 298.19 452.21C294.397 453.883 290.592 453.908 286.678 452.518C283.574 451.411 281.375 449.344 279.479 446.699C276.749 442.911 275.444 438.593 274.333 434.116C273.463 430.634 273.487 427.153 273.645 423.635C273.693 422.577 273.874 421.519 274.128 420.486C275.384 415.541 277.039 410.768 280.192 406.685C282.583 403.585 285.132 400.608 288.623 398.787C291.184 397.459 293.793 396.266 296.68 395.626C299.603 394.986 302.526 394.494 305.474 394.347C309.424 394.15 313.362 394.568 317.203 395.528C320.79 396.426 324.45 397.114 327.845 398.664C330.333 399.796 332.749 401.137 335.298 402.109C337.677 403.019 339.719 404.47 341.736 405.885C346.036 408.899 350.071 412.257 353.646 416.181C354.842 417.497 356.026 418.813 357.089 420.252C358.551 422.245 359.541 424.447 360.471 426.71C361.051 428.137 361.595 429.589 362.09 431.053C363.105 434.054 364.155 437.043 364.639 440.217C365.086 443.12 365.714 445.998 366.378 448.864C366.97 451.423 366.982 454.043 367.405 456.626C367.574 457.623 368.045 458.299 368.854 458.754C369.181 458.939 369.627 459.025 369.99 458.976C371.174 458.791 372.357 457.463 372.357 456.257C372.345 454.953 372.225 453.65 372.092 452.346C372.007 451.46 371.693 450.574 371.705 449.689C371.729 448.422 371.439 447.302 370.944 446.158C370.594 445.371 370.57 444.436 370.388 443.563C370.316 443.243 370.231 442.923 370.111 442.628C369.193 440.278 368.492 437.867 368.057 435.37C367.948 434.743 367.61 434.153 367.393 433.55C367.079 432.713 366.668 431.889 366.487 431.016C365.871 428.1 364.53 425.48 363.358 422.799C362.754 421.396 361.8 420.154 361.003 418.85C360.87 418.641 360.677 418.469 360.592 418.247C359.783 416.021 357.923 414.619 356.509 412.86C356.461 412.798 356.413 412.724 356.34 412.675C354.069 411.224 352.704 408.751 350.518 407.238C348.537 405.86 346.918 404.015 344.744 402.896C344.043 402.527 343.476 401.887 342.823 401.42C342.304 401.038 341.772 400.657 341.193 400.399C338.608 399.292 336.071 398.111 333.691 396.561C332.628 395.872 331.348 395.528 330.152 395.06C329.101 394.654 328.038 394.285 326.975 393.916C326.528 393.756 326.057 393.584 325.598 393.535C323.339 393.289 321.249 392.489 319.123 391.776C317.843 391.345 316.466 391.185 315.113 391.001C312.407 390.632 309.702 390.324 306.984 390.004C306.513 389.943 306.005 389.882 305.546 389.98C303.589 390.398 301.512 389.992 299.627 390.902C299.217 391.099 298.673 391.013 298.202 391.038C296.1 391.173 294.059 391.579 292.186 392.612C292.041 392.686 291.909 392.809 291.752 392.821C289.07 393.104 287.173 395.146 284.77 396.044C284.202 396.253 283.707 396.77 283.26 397.237C281.798 398.738 280.469 400.399 278.911 401.789C276.217 404.175 274.901 407.447 273.161 410.461C272.968 410.805 272.859 411.224 272.787 411.63C272.4 413.733 271.386 415.603 270.721 417.596C270.238 419.047 269.948 420.597 269.815 422.135C269.598 424.742 269.61 427.362 269.441 429.97C269.38 430.88 269.429 431.742 269.658 432.627C269.912 433.574 269.972 434.571 270.226 435.506C270.564 436.748 271.035 437.966 271.422 439.196C271.808 440.438 272.086 441.718 272.521 442.948C272.787 443.698 273.27 444.387 273.657 445.1C273.729 445.236 273.91 445.346 273.947 445.494C274.599 448.163 276.749 449.836 278.174 451.989C278.561 452.579 279.165 453.071 279.757 453.465C282.318 455.162 285.132 456.393 288.007 457.401C289.251 457.832 290.701 457.758 292.054 457.733C293.334 457.709 294.614 457.487 295.883 457.278C296.511 457.18 297.163 457.008 297.707 456.7C299.603 455.63 301.717 455.064 303.626 454.055C305.027 453.317 306.283 452.26 307.563 451.288C308.264 450.746 308.808 449.996 309.508 449.455C311.199 448.126 312.407 446.453 313.47 444.584C315.029 441.828 315.536 438.839 315.173 435.825C314.859 433.218 313.494 430.905 311.622 429.035C310.607 428.014 309.508 427.092 308.095 426.551C306.392 425.899 304.737 425.763 302.913 425.936C299.857 426.231 297.646 428.15 295.182 429.601C294.107 431.04 293.056 432.467 292.996 434.399C292.959 435.383 292.887 436.354 292.754 437.326C292.597 438.507 293.008 439.405 293.89 440.131C294.53 440.598 295.158 441.103 295.931 441.742Z"
-    fill="black"
-   />
-   <path
-    d="M699.007 276.869C699.669 277.146 700.224 277.537 700.814 277.612C702.527 277.827 704.239 277.915 705.964 278.041C706.176 278.053 706.46 278.003 706.601 277.864C707.499 276.957 708.727 277.562 709.755 276.819C710.085 277.877 710.451 278.721 710.605 279.616C710.759 280.422 710.345 281.178 709.424 281.481C707.015 282.262 704.605 281.985 702.196 281.67C701.18 281.531 700.188 281.304 699.172 281.191C698.806 281.153 698.428 281.279 698.05 281.342C697.696 282.98 697.554 284.593 696.976 286.017C696.397 287.454 696.787 289.155 695.617 290.49C694.436 290.276 693.066 290.339 692.145 288.84C692.393 287.403 693.078 286.067 693.48 284.656C693.87 283.295 694.602 281.884 694.141 280.246C690.61 279.83 687.315 278.268 683.831 278.494C683.205 278.003 683.228 277.423 683.323 276.831C683.394 276.415 683.476 276.012 683.606 275.609C684.315 273.416 685.756 272.686 687.823 273.53C688.98 274.009 690.031 274.727 691.224 275.143C692.358 275.533 693.492 275.533 694.826 275.924C694.72 272.144 695.204 268.653 694.082 265.176C695.594 264.029 697.188 264.382 698.594 265.049C699.539 265.503 699.798 266.763 699.657 267.897C699.562 268.628 699.196 269.334 699.137 270.065C698.948 272.307 698.948 274.55 699.007 276.869Z"
-    fill="black"
-   />
-   <path
-    d="M536.926 106.273C536.366 107.275 535.854 108.252 535.281 109.206C534.55 110.437 534.477 111.801 534.514 113.152C534.575 116.109 534.441 119.09 534.879 121.999C535.233 124.34 534.831 126.633 535.147 128.938C535.208 129.409 535.123 129.928 535.281 130.375C535.951 132.257 534.965 133.766 534.173 135.311C533.746 136.156 533.076 136.373 532.15 136.132C531.468 135.963 530.737 135.987 529.981 135.926C529.75 136.506 529.555 136.928 529.409 137.375C528.982 138.69 529.372 139.595 530.725 139.969C532.662 140.513 534.587 140.392 536.329 139.257C537.669 138.388 538.644 137.206 538.888 135.613C539.095 134.188 539.253 132.74 539.241 131.316C539.205 128.999 539.034 126.67 538.888 124.352C538.754 122.192 538.559 120.044 538.4 117.883C538.291 116.363 538.157 114.854 538.132 113.333C538.108 111.801 538.644 110.521 540.131 109.761C540.91 109.351 541.654 108.844 542.385 108.337C543.932 107.287 543.993 105.368 542.458 104.33C541.386 103.618 540.252 102.978 539.107 102.399C538.303 101.988 537.852 101.361 537.718 100.552C537.584 99.6954 537.548 98.8023 537.596 97.9333C537.791 94.5782 538.023 91.2351 538.266 87.8919C538.425 85.7316 538.498 83.5713 538.815 81.4351C539.168 78.9851 538.644 76.9092 537.182 74.8816C536.086 73.373 534.928 71.9609 533.539 70.7299C531.943 69.2937 530.432 69.8127 529.713 71.8885C529.348 72.9385 529.506 73.7592 530.201 74.5437C530.676 75.0868 531.102 75.7023 531.651 76.1609C532.552 76.9092 533.161 77.8023 533.429 78.9368C533.478 79.1661 533.783 79.3713 533.758 79.5644C533.466 82.123 534.526 84.6333 533.99 87.204C533.905 87.5902 533.783 88.0006 533.819 88.3747C534.112 91.6695 533.332 94.904 533.308 98.1868C533.283 101.252 534.063 104.028 536.743 105.959C536.853 106.02 536.877 106.201 536.926 106.273Z"
-    fill="black"
-   />
-   <path
-    d="M666.415 170.388C665.874 171.042 665.459 171.745 664.884 172.233C664.055 172.935 663.192 172.994 662.479 172.054C661.731 171.054 660.948 170.078 660.131 169.15C658.692 167.52 657.219 165.925 655.735 164.33C655.597 164.176 655.332 164.152 655.044 164.033C654.227 164.568 653.444 165.14 652.811 165.973C651.868 167.198 650.843 168.365 649.612 169.305C649.439 169.436 649.186 169.614 649.163 169.793C648.921 171.197 647.644 171.661 646.907 172.625C646.044 172.137 645.204 171.745 644.686 170.888C644.295 170.245 644.26 169.71 644.732 169.15C645.388 168.377 646.021 167.555 646.792 166.925C648.473 165.544 649.842 163.83 651.407 162.331C651.787 161.974 652.029 161.462 652.294 161.081C652.178 160.819 652.155 160.629 652.052 160.546C649.554 158.594 647.54 156.083 645.112 154.048C644.479 153.512 643.731 153.12 642.96 152.596C643.374 150.787 644.675 150.001 646.067 149.347C646.297 149.24 646.677 149.359 646.942 149.478C648.242 150.037 649.255 150.942 649.957 152.227C651.258 154.595 653.26 156.25 655.389 158.023C655.999 157.297 656.46 156.559 657.093 156.035C658.52 154.833 659.797 153.488 660.948 152.001C661.57 151.215 662.168 150.418 662.824 149.561C663.848 149.454 664.412 150.228 664.976 150.965C666.254 152.655 666.127 154.048 664.585 155.5C664.355 155.714 664.124 155.928 663.848 156.059C661.593 157.166 659.935 158.999 658.267 160.962C658.267 161.867 658.83 162.509 659.475 163.057C660.879 164.271 662.145 165.604 663.181 167.163C663.261 167.294 663.33 167.496 663.445 167.532C664.93 167.924 665.31 169.495 666.415 170.388Z"
-    fill="black"
-   />
-   <path
-    d="M740.026 356.911C738.569 355.641 736.94 355.888 735.481 355.524C734.011 355.155 732.585 354.657 731.123 354.265C729.656 353.885 728.17 353.563 726.619 353.206C726.53 351.706 727.219 350.391 728.011 349.15C728.516 348.377 729.926 348.134 730.63 348.807C731.995 350.085 733.648 350.555 735.386 350.923C736.173 351.085 736.963 351.315 737.667 351.685C739.337 352.555 741.121 352.977 742.975 353.004C745.044 353.038 745.984 354.631 745.608 356.485C745.361 357.701 744.972 358.882 744.645 360.07C744.438 360.866 744.158 361.651 744.036 362.462C743.852 363.752 743.697 364.999 743.039 366.22C742.523 367.173 742.635 368.458 742.29 369.549C741.5 371.985 741.043 374.455 740.716 376.995C740.461 379.022 739.766 380.992 739.311 383.003C739.066 384.101 738.92 385.207 738.69 386.52C740.178 384.584 741.811 382.973 742.789 381.035C744.149 378.347 746.375 376.307 747.693 373.631C748.017 372.979 748.501 372.343 749.059 371.903C750.614 370.672 751.665 368.968 753.064 367.593C753.64 367.027 754.125 366.313 754.511 365.591C755.645 363.456 757.219 361.68 758.699 359.807C759.319 359.033 759.826 358.182 760.404 357.381C760.787 356.854 761.156 356.297 761.595 355.829C763.456 353.849 764.944 351.612 766.285 349.245C767.098 347.828 768.189 346.571 769.174 345.252C770.264 343.811 771.396 342.397 772.474 340.952C773.365 339.758 774.24 338.572 775.577 337.832C776.33 337.419 777.073 337.225 777.901 337.493C778.438 337.676 778.989 337.784 779.514 338.002C781.462 338.769 783.434 339.427 785.417 340.088C787.218 340.687 788.977 341.482 790.753 342.191C792.776 342.997 794.825 343.615 796.965 344.08C798.749 344.463 800.456 345.188 802.199 345.767C803.412 346.18 804.612 346.706 805.849 347.008C807.736 347.466 809.515 348.163 811.18 349.162C811.518 349.369 811.906 349.54 812.285 349.59C813.704 349.771 815.181 349.708 816.481 350.465C817.369 350.216 818.032 350.783 818.728 351.138C820.033 351.804 821.328 352.389 822.838 352.377C823.375 352.376 823.903 352.661 824.483 352.832C824.56 353.357 824.664 353.838 824.692 354.32C824.765 355.645 824.413 356.116 823.19 356.449C822.205 356.718 821.253 356.551 820.338 356.24C816.39 354.898 812.431 353.552 808.496 352.175C806.372 351.439 804.272 350.634 802.152 349.886C799.802 349.062 797.442 348.233 795.065 347.49C791.998 346.526 788.901 345.684 785.826 344.743C782.979 343.88 780.16 342.974 777.371 342.092C776.416 342.423 775.835 343.118 775.292 343.853C773.294 346.574 771.304 349.311 769.314 352.047C768.435 353.245 767.567 354.446 766.699 355.648C764.583 358.604 762.49 361.569 760.356 364.506C758.8 366.642 757.376 368.915 755.563 370.793C753.435 372.997 751.725 375.513 749.874 377.917C747.867 380.516 746.17 383.366 744.326 386.087C743.038 387.999 741.746 389.922 740.397 391.787C739.048 393.652 737.627 395.467 736.028 397.129C735.689 397.486 735.377 397.879 735.02 398.217C734.382 398.853 733.642 399.114 732.809 398.712C731.975 398.311 731.71 397.551 731.78 396.669C731.823 396.093 731.961 395.536 732.104 394.967C732.667 392.56 733.276 390.167 733.805 387.748C734.191 385.975 734.641 384.238 735.061 382.477C735.855 379.203 736.531 375.901 736.958 372.541C737.43 368.789 737.798 365.002 739.096 361.374C739.63 359.954 739.736 358.44 740.026 356.911Z"
-    fill="black"
-   />
-   <path
-    d="M469.056 213.259C469.439 211.801 470.533 211.282 471.708 211.011C474.316 210.415 476.488 209.006 478.631 207.522C478.827 207.389 479.009 207.172 479.217 207.136C481.972 206.628 484.196 204.76 486.95 204.252C487.467 204.156 487.963 203.819 488.439 203.552C489.198 203.126 489.878 202.548 490.672 202.233C494.973 200.519 498.77 197.949 502.689 195.571C503.893 194.838 505.189 194.557 506.646 195.188C507.265 196.296 507.782 197.571 508.629 198.586C509.517 199.665 509.903 200.925 510.502 202.114C511.14 203.379 511.984 204.525 512.227 205.983C512.314 206.482 512.781 206.928 513.077 207.391C513.467 207.984 513.92 208.549 514.233 209.18C514.83 210.381 515.195 211.722 515.936 212.82C516.697 213.956 517.312 215.111 517.794 216.371C517.904 216.669 518.04 216.959 518.228 217.207C519.204 218.57 519.877 220.054 520.245 221.681C520.58 223.174 521.409 224.246 523.03 224.908C524.091 225.006 525.156 224.424 526.224 223.926C527.229 223.457 528.267 223.122 529.328 222.79C530.665 222.359 531.986 221.759 533.157 220.987C534.563 220.062 536.123 219.477 537.62 218.802C538.94 218.214 540.169 217.567 541.348 216.736C543.575 215.155 544.29 215.32 545.633 217.691C546.213 218.723 546.729 219.795 547.258 220.857C547.896 222.122 548.545 223.388 549.158 224.662C549.691 225.796 550.12 226.989 550.705 228.093C552.386 231.29 553.442 234.796 555.531 237.8C555.84 238.253 556.134 238.739 556.328 239.25C556.782 240.447 556.181 241.594 554.895 241.996C553.297 242.493 551.817 243.229 550.322 243.989C548.788 244.767 547.074 245.227 545.441 245.827C544.85 246.046 544.197 246.174 543.7 246.523C542.676 247.264 541.575 247.842 540.48 248.468C539.36 249.103 538.182 249.409 537.024 249.849C536.659 249.986 536.244 250.034 535.889 250.196C534.888 250.63 534.324 251.245 534.277 252.504C534.22 253.952 533.717 255.396 533.41 256.828C533.339 257.129 533.196 257.435 533.173 257.742C532.991 259.856 532.788 261.954 532.678 264.076C532.509 267.454 532.4 270.827 532.255 274.208C532.234 274.599 532.107 274.978 532.074 275.368C531.872 277.24 531.385 279.08 531.684 280.998C531.803 281.751 531.732 282.577 531.55 283.331C531.314 284.342 531.356 285.348 531.323 286.369C531.282 287.366 530.651 288.355 530.685 289.324C530.724 290.247 530.953 291.191 530.835 292.119C530.681 293.365 530.408 294.598 530.155 295.964C529.759 296.062 529.331 296.227 528.88 296.271C526.11 296.491 523.344 296.998 520.564 296.358C516.31 296.6 512.208 295.416 508.016 295.021C506.002 294.838 503.894 294.74 501.982 293.771C501.447 293.507 500.739 293.486 500.11 293.51C497.778 293.589 495.563 292.847 493.272 292.669C491.769 292.545 490.259 292.492 488.762 292.416C488.451 292.404 488.111 292.532 487.815 292.487C486.186 292.205 484.657 292.627 483.118 293.035C482.819 293.12 482.493 293.131 482.161 293.188C481.123 292.581 480.817 291.675 480.969 290.56C481.017 290.244 481.032 289.9 481.182 289.643C482.204 287.864 482.05 285.832 482.472 283.937C482.875 282.087 483.29 280.251 483.505 278.368C483.728 276.415 484.129 274.481 484.376 272.531C484.552 271.204 484.561 269.858 484.725 268.529C484.951 266.659 484.891 264.757 485.626 262.957C485.794 262.547 485.697 262.024 485.703 261.547C485.722 258.855 486.19 256.227 486.577 253.577C486.794 252.099 486.845 250.603 486.979 249.116C487.028 248.573 487.01 247.999 487.165 247.48C487.548 246.129 487.029 245.092 486.34 243.953C484.832 241.456 483.504 238.848 482.136 236.284C480.986 234.126 479.861 231.959 478.724 229.79C478.576 229.511 478.328 229.256 478.265 228.963C477.753 226.174 475.791 224.067 474.681 221.556C473.398 218.667 471.702 216.017 469.367 213.796C469.236 213.685 469.195 213.514 469.056 213.259ZM492.261 248.718C491.781 250.07 491.216 251.221 490.98 252.446C490.735 253.753 490.56 255.069 490.313 256.388C489.912 258.535 489.85 260.758 489.602 262.935C489.29 265.665 488.932 268.378 488.335 271.075C488.012 272.529 487.675 274.004 487.627 275.477C487.59 276.748 487.333 277.947 487.121 279.174C487.035 279.713 486.791 280.269 486.852 280.789C486.97 281.863 486.643 282.828 486.272 283.763C486.051 284.322 485.975 284.778 486.152 285.335C486.31 285.853 486.392 286.399 486.523 286.939C486.725 287.069 486.918 287.27 487.132 287.294C488.701 287.473 490.217 288.112 491.737 288.19C493.748 288.288 495.632 288.873 497.604 289.098C502.165 289.643 506.707 290.46 511.139 291.705C512.13 291.986 513.168 292.068 514.19 292.185C516.934 292.51 519.713 292.422 522.444 292.961C522.667 293.01 522.932 293.017 523.143 292.957C524.253 292.619 525.343 292.243 526.432 291.879C526.674 290.5 526.818 289.24 527.126 288.024C527.509 286.565 527.642 285.079 527.705 283.584C527.729 282.954 527.764 282.327 527.847 281.704C528.131 279.436 528.161 277.162 528.097 274.878C528.045 273.227 527.901 271.541 528.204 269.943C528.543 268.134 528.528 266.367 528.42 264.567C528.379 263.871 528.461 263.153 528.517 262.456C528.689 260.103 528.909 257.756 529.033 255.41C529.085 254.426 528.911 253.417 528.839 252.467C526.82 252.022 524.976 251.406 523.103 251.251C519.788 250.98 516.474 250.482 513.119 250.552C511.538 250.586 509.894 250.649 508.385 250.262C506.665 249.827 504.97 249.597 503.212 249.599C502.67 249.597 502.108 249.461 501.568 249.34C498.512 248.693 495.44 248.39 492.261 248.718ZM517.615 226.054C517.415 225.268 517.33 224.543 517.04 223.914C516.209 222.126 515.301 220.378 514.394 218.629C512.932 215.827 511.417 213.067 509.981 210.256C508.315 207.026 506.725 203.769 505.071 200.54C504.875 200.148 504.501 199.843 504.183 199.461C503.266 200.012 502.412 200.427 501.662 200.985C498.978 202.979 495.894 204.26 493 205.885C491.893 206.51 490.789 207.111 489.675 207.699C487.152 209.02 484.639 210.366 482.074 211.635C480.237 212.546 478.361 213.369 476.841 214.769C476.845 217.046 478.045 218.876 478.989 220.724C480.779 224.243 482.545 227.747 483.853 231.473C484.162 232.354 484.645 233.185 485.092 234.011C485.946 235.598 487.028 237.093 487.644 238.761C488.222 240.341 489.357 241.043 490.596 241.792C491.259 241.367 491.77 240.996 492.323 240.69C495.836 238.707 499.36 236.737 502.882 234.779C503.922 234.206 504.941 233.512 506.069 233.224C507.198 232.935 508.104 232.383 509.037 231.798C509.84 231.306 510.654 230.815 511.499 230.375C513.653 229.226 515.858 228.154 517.615 226.054ZM542.509 221.006C542.233 220.998 542.048 220.929 541.922 220.986C536.46 223.689 530.94 226.277 525.313 228.638C525.186 228.695 525.118 228.866 524.92 229.13C525.585 230.267 526.225 231.508 526.996 232.669C528.258 234.578 529.591 236.506 530.17 238.706C530.745 240.846 531.949 242.736 532.412 244.899C532.552 245.571 532.916 246.185 533.235 246.972C534.245 246.563 535.202 246.196 536.136 245.814C538.475 244.841 540.802 243.855 543.141 242.882C545.328 241.964 547.524 241.082 549.712 240.152C550.434 239.841 551.095 239.428 551.875 239.028C549.13 232.706 545.626 226.917 542.509 221.006ZM520.257 230.85C519.583 231.167 518.904 231.626 518.16 231.815C516.439 232.238 515.001 233.23 513.501 234.037C509.947 235.955 506.338 237.725 502.596 239.288C500.497 240.169 498.568 241.451 496.596 242.597C496.221 242.816 495.968 243.24 495.599 243.627C496.248 243.952 496.739 244.401 497.233 244.398C498.775 244.383 500.185 245.009 501.667 245.203C506.221 245.806 510.787 246.411 515.343 246.991C516.281 247.11 517.244 247.113 518.188 247.173C520.159 247.303 522.119 247.431 524.089 247.573C524.245 247.579 524.421 247.623 524.543 247.708C525.647 248.478 526.842 248.03 527.997 247.936C528.92 247.862 529.311 247.287 529.027 246.504C528.178 244.13 527.556 241.674 526.152 239.523C525.64 238.737 525.065 237.98 524.939 236.976C524.894 236.626 524.587 236.257 524.317 235.987C523.224 234.909 522.522 233.577 521.805 232.267C521.456 231.631 521.061 231.085 520.257 230.85Z"
-    fill="black"
-   />
-   <path
-    d="M223.573 290.854L246.014 243.829L317.564 303.049L223.573 290.854Z"
-    fill="white"
-   />
-   <path
-    d="M298.738 273.435C299.458 274.949 300.616 275.761 302.068 276.265C302.665 276.474 303.202 276.88 303.714 277.274C305.141 278.345 306.544 279.427 308.251 280.043C308.775 280.227 309.275 280.51 309.739 280.83C311.605 282.122 313.751 282.934 315.398 284.595C315.934 285.137 316.788 285.432 317.544 285.69C319.02 286.195 320.288 287.044 321.618 287.819C322.666 288.434 323.801 288.889 324.886 289.431C325.398 289.677 325.91 289.923 326.41 290.206C328.142 291.227 328.203 293.27 326.508 294.365C326.118 294.623 325.593 294.857 325.154 294.82C323.544 294.66 321.947 295.374 320.325 294.796C319.91 294.648 319.374 294.832 318.886 294.894C318.166 294.968 317.434 295.165 316.715 295.14C313.971 295.054 311.239 295.054 308.495 295.202C307.471 295.263 306.422 294.931 305.385 294.808C304.739 294.734 304.068 294.623 303.446 294.709C301.092 295.054 298.763 294.673 296.421 294.66C296.336 294.66 296.25 294.697 296.177 294.673C292.836 293.799 289.409 294.537 286.03 294.193C283.945 293.983 281.811 294.266 279.713 294.168C276.323 293.996 272.932 294.033 269.542 293.7C266.176 293.368 262.785 293.024 259.358 293.282C257.444 293.43 255.492 293.085 253.553 292.987C251.797 292.901 250.102 292.716 248.346 293.602C246.858 294.353 245.15 294.057 243.967 292.421C243.687 292.445 243.37 292.421 243.065 292.495C240.784 293.048 238.479 293.134 236.15 292.827C233.881 292.532 233.881 292.544 231.967 293.577C231.503 293.491 231.015 293.417 230.552 293.294C228.918 292.864 228.284 291.683 229.015 290.145C230.028 288.004 231.162 285.924 232.235 283.82C232.454 283.389 232.698 282.959 232.906 282.516C234.235 279.71 235.564 276.905 236.857 274.087C237.272 273.201 237.54 272.242 237.943 271.356C238.723 269.67 239.199 267.812 240.54 266.409C240.796 266.139 240.857 265.659 240.979 265.277C241.857 262.546 243.114 259.999 244.528 257.513C245.126 256.455 245.577 255.311 246.004 254.167C246.723 252.186 247.894 250.389 248.309 248.261C248.529 247.104 249.455 246.747 250.553 247.165C250.931 247.313 251.309 247.498 251.602 247.768C253.7 249.688 256.334 250.561 258.907 251.509C259.456 251.718 259.944 251.915 260.334 252.346C260.944 253.035 261.688 253.441 262.566 253.724C263.407 253.994 264.212 254.437 264.993 254.856C265.481 255.126 266.103 255.409 266.322 255.865C267.017 257.267 268.334 257.673 269.566 258.264C269.932 258.436 270.322 258.572 270.639 258.805C274.2 261.414 278.262 263.1 282.03 265.314C283.274 266.053 284.664 266.545 285.811 267.455C286.006 267.603 286.226 267.726 286.445 267.812C288.591 268.587 290.543 269.707 292.262 271.22C292.494 271.43 292.836 271.516 293.128 271.651C294.226 272.168 295.409 272.562 296.409 273.226C297.165 273.694 297.811 273.497 298.738 273.435ZM292.75 276.585C292.482 275.822 291.726 275.588 291.128 275.219C290.092 274.592 288.945 274.124 287.945 273.435C285.201 271.528 282.237 270.039 279.237 268.649C275.347 266.828 271.713 264.551 268.017 262.386C266.81 261.685 265.59 261.106 264.31 260.713C262.7 262.312 261.2 263.912 259.59 265.388C258.456 266.434 257.224 267.381 255.504 267.209C255.041 267.16 254.553 267.357 254.078 267.492C252.956 267.812 251.834 267.738 250.712 267.468C248.992 267.049 247.26 266.68 245.223 266.225C243.784 269.141 242.187 271.996 240.967 275.133C242.162 275.982 243.284 276.77 244.394 277.545C246.102 278.763 247.431 280.375 248.028 282.344C248.565 284.103 249.053 285.949 249.004 287.856C248.992 288.237 249.126 288.606 249.163 288.877C249.394 289 249.468 289.074 249.529 289.074C254.553 289.062 259.553 289.615 264.566 289.579C268.688 289.554 272.822 289.689 276.945 289.788C278.64 289.825 280.335 289.923 282.018 290.009C283.799 290.095 285.579 290.218 287.36 290.28C287.799 290.292 288.25 290.095 288.652 289.997C289.396 282.467 289.396 282.122 292.75 276.585ZM292.836 290.292C294.275 290.378 295.482 290.489 296.689 290.514C299.275 290.575 301.86 290.6 304.446 290.637C308.324 290.686 312.202 290.735 316.081 290.76C316.532 290.76 317.056 290.87 317.361 290.354C317.349 290.268 317.361 290.145 317.325 290.12C314.325 288.36 311.324 286.601 308.312 284.866C306.763 283.968 305.092 283.279 303.69 282.196C302.373 281.187 300.897 280.535 299.47 279.821C298.104 279.144 295.689 279.255 294.909 281.224C294.458 282.356 294.019 283.512 293.421 284.571C292.433 286.342 292.604 288.164 292.836 290.292ZM247.858 261.094C250.407 263.136 251.724 263.616 254.151 263.21C256.907 262.743 259.114 261.426 260.505 258.682C259.2 257.907 257.895 257.132 256.59 256.369C255.822 255.926 255.151 255.348 254.212 255.212C253.431 255.102 252.675 254.548 251.919 254.573C250.833 254.61 250.992 255.963 250.333 256.554C249.016 257.747 248.394 259.334 247.858 261.094ZM245.089 288.582C245.187 287.413 245.394 286.367 244.894 285.37C244.272 284.152 243.809 282.713 242.845 281.851C241.76 280.879 240.65 279.624 238.808 279.587C238.491 280.067 238.126 280.596 237.796 281.138C236.65 282.983 236.296 285.235 234.772 286.896C234.369 287.327 234.442 287.967 235.077 288.606C238.272 288.385 241.626 288.791 245.089 288.582Z"
-    fill="black"
-   />
-   <path
-    d="M528.144 348.926C528.556 348.171 528.841 347.443 529.299 346.842C530.182 345.673 531.277 345.579 532.445 346.519C534.856 348.44 537.298 350.257 539.95 351.813C541.147 352.515 542.158 353.53 543.229 354.45C544.819 355.82 546.39 357.221 547.96 358.623C548.286 358.909 548.666 359.197 548.866 359.578C550.129 361.917 551.664 364.158 552.509 366.653C554.45 372.396 554.297 378.226 551.92 383.774C550.54 386.998 548.762 390.032 546.644 392.839C543.436 397.089 540.374 401.472 537.989 406.285C536.786 408.707 535.849 411.319 535.126 413.941C534.541 416.083 534.775 418.325 535.873 420.385C537.139 422.738 538.797 424.803 540.431 426.886C540.96 427.566 541.698 428.106 542.382 428.644C543.896 429.839 545.417 431.005 546.961 432.153C549.356 433.94 551.868 435.457 554.797 436.122C555.214 436.211 555.603 436.415 555.934 436.537C556.475 437.528 556.274 438.402 555.796 439.158C555.337 439.882 554.748 440.552 554.111 441.122C553.842 441.356 553.285 441.364 552.864 441.317C550.758 441.092 548.811 440.302 547.373 438.709C546.018 437.194 544.277 436.184 542.829 434.797C541.011 433.039 538.931 431.554 536.985 429.945C536.791 429.781 536.627 429.57 536.463 429.359C535.133 427.525 533.755 425.715 532.507 423.823C531.43 422.181 531.107 420.271 530.677 418.357C530.285 416.625 530.289 415.001 530.731 413.38C531.221 411.613 531.912 409.912 532.531 408.185C532.612 407.937 532.877 407.744 532.906 407.506C533.227 404.997 534.553 402.903 535.764 400.834C536.898 398.903 537.967 396.918 539.477 395.195C541.114 393.322 542.544 391.22 544.019 389.204C545.163 387.64 546.447 385.978 546.746 383.869C546.788 383.628 546.963 383.386 547.118 383.176C548.345 381.499 548.569 379.474 548.95 377.498C549.115 376.644 549.265 375.781 549.346 374.904C549.403 374.292 549.353 373.675 549.303 373.058C549.252 372.441 549.072 371.838 549.06 371.213C548.975 368.353 547.866 365.927 546.072 363.781C544.83 362.297 543.655 360.758 542.412 359.274C541.508 358.196 540.63 357.057 539.24 356.573C537.672 354.871 535.241 354.428 533.734 352.632C533.575 352.447 533.302 352.286 533.075 352.28C531.532 352.209 530.576 351.265 529.693 350.1C529.374 349.595 528.766 349.355 528.144 348.926Z"
-    fill="black"
-   />
-   <path
-    d="M215.32 189.257C214.941 192.709 213.975 195.667 211.241 197.658C210.178 198.44 209.126 199.177 207.846 199.591C205.633 200.305 203.561 199.994 201.554 198.947C200.925 198.613 200.241 198.176 199.851 197.589C198.592 195.701 197.638 193.63 197.66 191.248C197.67 190.442 197.952 189.613 198.202 188.831C198.896 186.633 200.35 184.999 202.086 183.767C203.637 182.674 205.427 181.972 207.412 182.49C207.9 182.616 208.421 182.743 208.909 182.674C210.515 182.444 211.675 183.445 212.858 184.262C213.899 184.976 214.594 186.115 214.995 187.369C215.201 188.025 215.234 188.727 215.32 189.257ZM203.724 187.381C203.485 187.68 203.105 188.232 202.65 188.716C201.033 190.431 201.163 193.526 202.932 195.011C205.112 196.841 207.727 196.599 209.658 194.378C210.092 193.872 210.526 193.365 210.894 192.801C211.166 192.375 211.393 191.869 211.524 191.374C211.719 190.569 211.806 189.729 211.968 188.75C210.84 187.864 209.343 188.336 208.095 187.772C206.859 187.197 205.449 187.024 203.724 187.381Z"
-    fill="black"
-   />
-   <path
-    d="M451.053 492.521C450.29 493.395 450.379 494.721 451.252 495.484C452.126 496.247 453.452 496.157 454.215 495.284L451.053 492.521ZM413.283 275.967C412.144 275.749 411.044 276.496 410.826 277.635L407.272 296.193C407.054 297.332 407.801 298.432 408.94 298.65C410.078 298.868 411.178 298.121 411.396 296.982L414.555 280.486L431.051 283.645C432.19 283.863 433.29 283.117 433.508 281.978C433.726 280.839 432.98 279.739 431.841 279.521L413.283 275.967ZM413.748 283.128C414.378 284.101 415.678 284.38 416.652 283.75C417.625 283.12 417.903 281.82 417.273 280.846L413.748 283.128ZM422.341 289.036C421.751 288.039 420.463 287.709 419.465 288.299C418.467 288.89 418.137 290.178 418.728 291.176L422.341 289.036ZM423.4 299.404C423.955 300.422 425.23 300.798 426.249 300.243C427.267 299.688 427.642 298.413 427.088 297.395L423.4 299.404ZM431.54 305.898C431.019 304.863 429.757 304.446 428.721 304.967C427.685 305.488 427.268 306.75 427.79 307.786L431.54 305.898ZM431.919 316.311C432.408 317.362 433.657 317.817 434.708 317.327C435.759 316.838 436.214 315.589 435.725 314.538L431.919 316.311ZM439.654 323.295C439.194 322.23 437.959 321.74 436.894 322.2C435.83 322.66 435.34 323.895 435.799 324.96L439.654 323.295ZM439.441 333.715C439.871 334.792 441.093 335.315 442.17 334.884C443.246 334.453 443.77 333.232 443.339 332.155L439.441 333.715ZM446.786 341.108C446.384 340.02 445.176 339.465 444.088 339.868C443.001 340.27 442.446 341.478 442.848 342.565L446.786 341.108ZM446.026 351.5C446.4 352.597 447.594 353.183 448.691 352.809C449.788 352.434 450.374 351.241 450 350.144L446.026 351.5ZM452.982 359.26C452.636 358.153 451.458 357.537 450.351 357.883C449.245 358.23 448.629 359.408 448.975 360.514L452.982 359.26ZM451.693 369.602C452.011 370.717 453.173 371.363 454.288 371.045C455.403 370.727 456.049 369.565 455.731 368.45L451.693 369.602ZM458.242 377.709C457.953 376.587 456.809 375.91 455.686 376.199C454.563 376.488 453.887 377.632 454.175 378.755L458.242 377.709ZM456.412 387.967C456.67 389.097 457.795 389.804 458.926 389.546C460.056 389.288 460.763 388.162 460.505 387.032L456.412 387.967ZM462.507 396.412C462.282 395.274 461.177 394.535 460.04 394.761C458.902 394.986 458.163 396.091 458.389 397.229L462.507 396.412ZM460.092 406.546C460.282 407.69 461.364 408.463 462.508 408.272C463.651 408.081 464.424 407 464.233 405.856L460.092 406.546ZM465.658 415.369C465.506 414.219 464.451 413.411 463.302 413.563C462.152 413.715 461.344 414.77 461.496 415.92L465.658 415.369ZM462.566 425.319C462.675 426.474 463.7 427.321 464.854 427.212C466.008 427.102 466.856 426.078 466.746 424.924L462.566 425.319ZM467.452 434.504C467.392 433.346 466.405 432.456 465.247 432.516C464.089 432.576 463.199 433.563 463.259 434.721L467.452 434.504ZM463.517 444.17C463.519 445.33 464.461 446.268 465.62 446.266C466.78 446.265 467.718 445.323 467.716 444.164L463.517 444.17ZM467.448 453.815C467.517 452.658 466.634 451.664 465.477 451.595C464.319 451.527 463.325 452.409 463.257 453.567L467.448 453.815ZM462.358 462.908C462.201 464.057 463.005 465.115 464.153 465.272C465.302 465.43 466.361 464.626 466.518 463.477L462.358 462.908ZM464.741 473.03C465.013 471.903 464.32 470.768 463.193 470.496C462.066 470.224 460.931 470.917 460.659 472.044L464.741 473.03ZM457.9 480.868C457.472 481.946 457.998 483.166 459.076 483.594C460.154 484.022 461.374 483.495 461.802 482.418L457.9 480.868ZM457.26 491.267C457.899 490.299 457.631 488.997 456.663 488.358C455.695 487.72 454.393 487.987 453.755 488.955L457.26 491.267ZM411.151 279.208C412.024 280.495 412.89 281.802 413.748 283.128L417.273 280.846C416.399 279.495 415.516 278.163 414.625 276.851L411.151 279.208ZM418.728 291.176C420.316 293.858 421.874 296.603 423.4 299.404L427.088 297.395C425.539 294.552 423.956 291.763 422.341 289.036L418.728 291.176ZM427.79 307.786C429.198 310.585 430.575 313.429 431.919 316.311L435.725 314.538C434.364 311.618 432.969 308.736 431.54 305.898L427.79 307.786ZM435.799 324.96C437.048 327.85 438.262 330.771 439.441 333.715L443.339 332.155C442.147 329.176 440.918 326.22 439.654 323.295L435.799 324.96ZM442.848 342.565C443.945 345.528 445.005 348.508 446.026 351.5L450 350.144C448.967 347.118 447.895 344.104 446.786 341.108L442.848 342.565ZM448.975 360.514C449.923 363.54 450.829 366.571 451.693 369.602L455.731 368.45C454.857 365.385 453.94 362.32 452.982 359.26L448.975 360.514ZM454.175 378.755C454.967 381.835 455.713 384.908 456.412 387.967L460.505 387.032C459.798 383.936 459.043 380.826 458.242 377.709L454.175 378.755ZM458.389 397.229C459.01 400.36 459.578 403.468 460.092 406.546L464.233 405.856C463.713 402.734 463.137 399.583 462.507 396.412L458.389 397.229ZM461.496 415.92C461.917 419.098 462.274 422.234 462.566 425.319L466.746 424.924C466.449 421.784 466.086 418.596 465.658 415.369L461.496 415.92ZM463.259 434.721C463.426 437.945 463.513 441.098 463.517 444.17L467.716 444.164C467.712 441.017 467.622 437.793 467.452 434.504L463.259 434.721ZM463.257 453.567C463.065 456.804 462.766 459.923 462.358 462.908L466.518 463.477C466.942 460.376 467.251 457.15 467.448 453.815L463.257 453.567ZM460.659 472.044C459.895 475.208 458.976 478.156 457.9 480.868L461.802 482.418C462.961 479.499 463.937 476.36 464.741 473.03L460.659 472.044ZM453.755 488.955C452.909 490.238 452.008 491.428 451.053 492.521L454.215 495.284C455.303 494.039 456.317 492.697 457.26 491.267L453.755 488.955Z"
-    fill="black"
-   />
-   <path
-    d="M636.591 495.742C636.306 496.866 636.987 498.008 638.111 498.292C639.236 498.576 640.377 497.895 640.662 496.771L636.591 495.742ZM898.891 440.856C897.862 440.32 896.595 440.72 896.059 441.748L887.335 458.509C886.799 459.537 887.199 460.805 888.228 461.341C889.256 461.876 890.524 461.476 891.059 460.448L898.814 445.55L913.713 453.305C914.741 453.84 916.009 453.44 916.544 452.412C917.08 451.383 916.68 450.116 915.652 449.58L898.891 440.856ZM763.466 478.396L761.485 479.092L763.466 478.396ZM897.188 447.81C897.472 448.934 898.614 449.615 899.738 449.33C900.862 449.046 901.543 447.904 901.258 446.78L897.188 447.81ZM903.017 456.223C902.882 455.072 901.839 454.247 900.687 454.382C899.536 454.517 898.711 455.56 898.846 456.711L903.017 456.223ZM899.21 465.721C899.163 466.88 900.064 467.857 901.223 467.903C902.382 467.95 903.359 467.048 903.405 465.89L899.21 465.721ZM902.154 475.554C902.407 474.423 901.696 473.3 900.564 473.047C899.433 472.793 898.31 473.505 898.056 474.636L902.154 475.554ZM895.276 483.122C894.81 484.184 895.293 485.422 896.354 485.888C897.416 486.355 898.655 485.872 899.121 484.811L895.276 483.122ZM894.347 493.324C895.009 492.371 894.772 491.063 893.82 490.402C892.868 489.741 891.559 489.977 890.898 490.929L894.347 493.324ZM885.129 497.805C884.308 498.625 884.307 499.954 885.126 500.775C885.946 501.595 887.275 501.596 888.096 500.777L885.129 497.805ZM880.7 507.046C881.64 506.368 881.852 505.056 881.174 504.115C880.495 503.175 879.183 502.963 878.243 503.641L880.7 507.046ZM870.532 508.406C869.506 508.947 869.113 510.217 869.654 511.242C870.194 512.268 871.464 512.661 872.49 512.12L870.532 508.406ZM863.706 516.065C864.791 515.655 865.338 514.443 864.928 513.359C864.518 512.274 863.307 511.727 862.222 512.137L863.706 516.065ZM853.514 514.875C852.39 515.161 851.711 516.304 851.998 517.428C852.284 518.551 853.427 519.23 854.551 518.944L853.514 514.875ZM845.15 520.82C846.297 520.652 847.091 519.586 846.923 518.438C846.755 517.291 845.689 516.497 844.542 516.665L845.15 520.82ZM835.425 517.536C834.266 517.589 833.37 518.571 833.423 519.729C833.475 520.887 834.457 521.784 835.615 521.731L835.425 517.536ZM826.039 521.69C827.197 521.753 828.187 520.865 828.249 519.707C828.312 518.55 827.424 517.56 826.266 517.497L826.039 521.69ZM817.166 516.54C816.021 516.36 814.946 517.142 814.766 518.287C814.585 519.433 815.368 520.508 816.513 520.688L817.166 516.54ZM807.126 518.69C808.245 518.993 809.398 518.331 809.701 517.212C810.004 516.093 809.343 514.94 808.223 514.637L807.126 518.69ZM799.578 511.751C798.502 511.319 797.28 511.841 796.848 512.917C796.416 513.993 796.938 515.216 798.014 515.648L799.578 511.751ZM789.316 511.494C790.327 512.062 791.607 511.702 792.174 510.691C792.741 509.68 792.382 508.4 791.371 507.833L789.316 511.494ZM783.805 502.854C782.886 502.147 781.568 502.32 780.861 503.239C780.155 504.159 780.327 505.477 781.247 506.183L783.805 502.854ZM774.048 499.701C774.845 500.543 776.174 500.58 777.016 499.783C777.858 498.986 777.895 497.657 777.098 496.815L774.048 499.701ZM771.473 489.78C770.828 488.816 769.524 488.558 768.56 489.203C767.597 489.849 767.339 491.153 767.984 492.116L771.473 489.78ZM763.291 483.627C763.764 484.686 765.006 485.16 766.064 484.687C767.123 484.213 767.597 482.971 767.124 481.913L763.291 483.627ZM763.671 473.053C763.227 471.982 761.999 471.474 760.927 471.918C759.856 472.362 759.348 473.591 759.792 474.662L763.671 473.053ZM755.684 466.232C756.257 467.24 757.539 467.592 758.547 467.019C759.555 466.446 759.907 465.164 759.333 464.156L755.684 466.232ZM753.83 455.896C753.118 454.981 751.799 454.816 750.884 455.528C749.969 456.24 749.804 457.559 750.516 458.474L753.83 455.896ZM744.23 451.612C745.082 452.398 746.41 452.346 747.197 451.494C747.984 450.643 747.931 449.314 747.079 448.528L744.23 451.612ZM739.161 442.409C738.183 441.785 736.885 442.072 736.261 443.05C735.638 444.027 735.924 445.325 736.902 445.949L739.161 442.409ZM728.685 441.713C729.76 442.147 730.983 441.627 731.417 440.552C731.851 439.476 731.331 438.253 730.256 437.819L728.685 441.713ZM720.664 434.938C719.528 434.706 718.419 435.439 718.187 436.575C717.956 437.711 718.689 438.82 719.825 439.052L720.664 434.938ZM710.613 438.002C711.772 438.034 712.738 437.121 712.77 435.962C712.802 434.803 711.888 433.837 710.729 433.805L710.613 438.002ZM700.766 434.347C699.617 434.502 698.811 435.56 698.967 436.709C699.122 437.858 700.179 438.664 701.328 438.508L700.766 434.347ZM692.209 440.456C693.322 440.131 693.96 438.965 693.634 437.852C693.309 436.739 692.142 436.101 691.03 436.427L692.209 440.456ZM681.714 439.888C680.658 440.366 680.189 441.61 680.668 442.666C681.146 443.723 682.39 444.191 683.447 443.712L681.714 439.888ZM675.193 448.14C676.176 447.526 676.475 446.231 675.861 445.248C675.247 444.264 673.951 443.965 672.968 444.579L675.193 448.14ZM664.915 450.358C664.017 451.092 663.885 452.415 664.619 453.312C665.353 454.21 666.676 454.342 667.574 453.608L664.915 450.358ZM660.696 459.984C661.496 459.146 661.465 457.817 660.626 457.016C659.787 456.216 658.458 456.247 657.658 457.086L660.696 459.984ZM651.271 464.642C650.577 465.571 650.767 466.887 651.696 467.581C652.625 468.275 653.94 468.085 654.635 467.156L651.271 464.642ZM649.455 475.009C650.036 474.006 649.694 472.721 648.69 472.14C647.687 471.559 646.402 471.902 645.821 472.905L649.455 475.009ZM641.371 481.718C640.908 482.781 641.394 484.019 642.457 484.482C643.52 484.945 644.757 484.459 645.22 483.396L641.371 481.718ZM641.942 492.221C642.286 491.114 641.667 489.937 640.559 489.594C639.452 489.25 638.276 489.869 637.932 490.976L641.942 492.221ZM895.919 443.349C896.395 444.859 896.818 446.346 897.188 447.81L901.258 446.78C900.868 445.236 900.424 443.672 899.924 442.086L895.919 443.349ZM898.846 456.711C899.212 459.837 899.326 462.84 899.21 465.721L903.405 465.89C903.53 462.781 903.407 459.558 903.017 456.223L898.846 456.711ZM898.056 474.636C897.388 477.617 896.452 480.445 895.276 483.122L899.121 484.811C900.405 481.887 901.426 478.8 902.154 475.554L898.056 474.636ZM890.898 490.929C889.195 493.382 887.262 495.675 885.129 497.805L888.096 500.777C890.398 498.478 892.494 495.993 894.347 493.324L890.898 490.929ZM878.243 503.641C875.823 505.387 873.243 506.977 870.532 508.406L872.49 512.12C875.365 510.605 878.112 508.913 880.7 507.046L878.243 503.641ZM862.222 512.137C859.404 513.202 856.492 514.116 853.514 514.875L854.551 518.944C857.676 518.147 860.737 517.187 863.706 516.065L862.222 512.137ZM844.542 516.665C841.532 517.106 838.485 517.397 835.425 517.536L835.615 521.731C838.814 521.586 842.001 521.281 845.15 520.82L844.542 516.665ZM826.266 517.497C823.212 517.332 820.171 517.014 817.166 516.54L816.513 520.688C819.661 521.184 822.845 521.517 826.039 521.69L826.266 517.497ZM808.223 514.637C805.268 513.837 802.378 512.875 799.578 511.751L798.014 515.648C800.972 516.835 804.019 517.848 807.126 518.69L808.223 514.637ZM791.371 507.833C788.718 506.344 786.188 504.685 783.805 502.854L781.247 506.183C783.797 508.143 786.496 509.912 789.316 511.494L791.371 507.833ZM777.098 496.815C775.044 494.644 773.16 492.3 771.473 489.78L767.984 492.116C769.805 494.837 771.837 497.364 774.048 499.701L777.098 496.815ZM767.124 481.913C766.514 480.55 765.954 479.146 765.447 477.701L761.485 479.092C762.031 480.647 762.634 482.158 763.291 483.627L767.124 481.913ZM765.447 477.701C764.887 476.106 764.295 474.557 763.671 473.053L759.792 474.662C760.386 476.092 760.95 477.569 761.485 479.092L765.447 477.701ZM759.333 464.156C757.638 461.176 755.798 458.424 753.83 455.896L750.516 458.474C752.356 460.838 754.084 463.422 755.684 466.232L759.333 464.156ZM747.079 448.528C744.567 446.206 741.919 444.169 739.161 442.409L736.902 445.949C739.449 447.574 741.899 449.458 744.23 451.612L747.079 448.528ZM730.256 437.819C727.135 436.56 723.927 435.603 720.664 434.938L719.825 439.052C722.843 439.667 725.805 440.551 728.685 441.713L730.256 437.819ZM710.729 433.805C707.419 433.714 704.087 433.898 700.766 434.347L701.328 438.508C704.432 438.089 707.536 437.918 710.613 438.002L710.729 433.805ZM691.03 436.427C687.88 437.348 684.766 438.505 681.714 439.888L683.447 443.712C686.323 442.41 689.252 441.322 692.209 440.456L691.03 436.427ZM672.968 444.579C670.196 446.311 667.504 448.239 664.915 450.358L667.574 453.608C670.026 451.601 672.573 449.777 675.193 448.14L672.968 444.579ZM657.658 457.086C655.409 459.443 653.274 461.964 651.271 464.642L654.635 467.156C656.536 464.612 658.563 462.22 660.696 459.984L657.658 457.086ZM645.821 472.905C644.197 475.711 642.708 478.651 641.371 481.718L645.22 483.396C646.494 480.474 647.911 477.677 649.455 475.009L645.821 472.905ZM637.932 490.976C637.447 492.538 636.999 494.126 636.591 495.742L640.662 496.771C641.052 495.228 641.479 493.711 641.942 492.221L637.932 490.976Z"
-    fill="black"
-   />
-   <path
-    d="M402.818 495.922C402.055 496.796 402.144 498.122 403.017 498.885C403.891 499.648 405.217 499.558 405.98 498.685L402.818 495.922ZM87.3964 266.052C86.7644 265.08 85.464 264.804 84.4918 265.436L68.6502 275.735C67.6781 276.367 67.4024 277.668 68.0344 278.64C68.6665 279.612 69.9669 279.888 70.939 279.256L85.0204 270.101L94.1754 284.182C94.8074 285.154 96.1079 285.43 97.08 284.798C98.0521 284.166 98.3278 282.865 97.6958 281.893L87.3964 266.052ZM363.482 350.803L362.722 352.76L363.482 350.803ZM137.152 389.651L136.92 391.738L136.97 391.743L137.02 391.746L137.152 389.651ZM409.14 494.578C409.799 493.625 409.56 492.317 408.607 491.658C407.653 490.998 406.345 491.237 405.686 492.191L409.14 494.578ZM410.63 483.87C410.096 484.9 410.497 486.167 411.526 486.701C412.555 487.235 413.823 486.834 414.357 485.805L410.63 483.87ZM418.615 476.575C419.054 475.502 418.54 474.276 417.467 473.837C416.393 473.398 415.168 473.912 414.729 474.985L418.615 476.575ZM418.057 465.788C417.707 466.894 418.319 468.074 419.425 468.424C420.53 468.774 421.71 468.162 422.06 467.057L418.057 465.788ZM424.724 457.291C424.984 456.161 424.277 455.035 423.147 454.776C422.017 454.517 420.891 455.223 420.632 456.353L424.724 457.291ZM422.409 446.748C422.248 447.897 423.048 448.958 424.196 449.12C425.345 449.281 426.406 448.481 426.568 447.333L422.409 446.748ZM427.517 437.236C427.571 436.078 426.676 435.095 425.518 435.041C424.359 434.987 423.377 435.882 423.323 437.041L427.517 437.236ZM423.285 427.308C423.35 428.466 424.341 429.352 425.499 429.287C426.656 429.222 427.542 428.231 427.477 427.074L423.285 427.308ZM426.344 416.952C426.15 415.809 425.065 415.041 423.922 415.235C422.779 415.43 422.01 416.514 422.205 417.657L426.344 416.952ZM420.003 408.216C420.336 409.327 421.506 409.958 422.617 409.625C423.727 409.292 424.358 408.122 424.025 407.011L420.003 408.216ZM420.473 397.439C420 396.381 418.759 395.906 417.7 396.379C416.641 396.853 416.167 398.094 416.64 399.153L420.473 397.439ZM412.126 390.612C412.734 391.599 414.027 391.906 415.014 391.298C416.002 390.69 416.309 389.397 415.701 388.41L412.126 390.612ZM409.78 380.063C409.05 379.163 407.728 379.024 406.827 379.754C405.926 380.484 405.788 381.806 406.518 382.707L409.78 380.063ZM399.987 375.585C400.82 376.391 402.149 376.37 402.956 375.537C403.762 374.704 403.741 373.375 402.908 372.568L399.987 375.585ZM395.204 365.9C394.287 365.189 392.969 365.356 392.258 366.272C391.548 367.189 391.714 368.508 392.631 369.218L395.204 365.9ZM384.639 363.634C385.62 364.252 386.916 363.958 387.534 362.977C388.152 361.996 387.858 360.7 386.877 360.082L384.639 363.634ZM378.111 355.074C377.081 354.541 375.815 354.944 375.282 355.974C374.749 357.003 375.151 358.27 376.181 358.803L378.111 355.074ZM367.33 354.638C368.396 355.093 369.63 354.598 370.086 353.532C370.542 352.466 370.047 351.232 368.981 350.777L367.33 354.638ZM359.361 347.364C358.231 347.104 357.104 347.809 356.843 348.939C356.583 350.069 357.288 351.196 358.418 351.456L359.361 347.364ZM349.24 350.289C350.398 350.341 351.38 349.445 351.433 348.287C351.485 347.128 350.589 346.147 349.431 346.094L349.24 350.289ZM339.574 346.2C338.416 346.268 337.533 347.261 337.601 348.419C337.669 349.576 338.662 350.459 339.819 350.392L339.574 346.2ZM330.379 351.28C331.53 351.136 332.345 350.087 332.201 348.936C332.057 347.786 331.008 346.97 329.857 347.114L330.379 351.28ZM320.204 348.557C319.061 348.753 318.293 349.838 318.489 350.981C318.685 352.124 319.77 352.892 320.913 352.696L320.204 348.557ZM311.508 354.47C312.644 354.238 313.377 353.129 313.145 351.993C312.913 350.857 311.804 350.124 310.668 350.356L311.508 354.47ZM301.204 352.411C300.074 352.669 299.367 353.795 299.625 354.925C299.884 356.056 301.01 356.763 302.14 356.504L301.204 352.411ZM292.785 358.732C293.911 358.454 294.598 357.316 294.321 356.191C294.043 355.065 292.905 354.377 291.779 354.655L292.785 358.732ZM282.393 357.035C281.271 357.327 280.598 358.473 280.889 359.595C281.181 360.718 282.327 361.391 283.449 361.099L282.393 357.035ZM274.146 363.562C275.265 363.261 275.929 362.109 275.628 360.99C275.326 359.87 274.174 359.206 273.055 359.508L274.146 363.562ZM263.742 362.043C262.624 362.35 261.967 363.506 262.275 364.624C262.582 365.742 263.737 366.399 264.855 366.091L263.742 362.043ZM255.566 368.661C256.683 368.351 257.337 367.193 257.027 366.076C256.716 364.959 255.559 364.305 254.442 364.615L255.566 368.661ZM245.148 367.2C244.031 367.511 243.377 368.668 243.688 369.786C243.998 370.903 245.156 371.557 246.273 371.246L245.148 367.2ZM236.972 373.823C238.09 373.514 238.746 372.359 238.438 371.241C238.13 370.123 236.974 369.466 235.856 369.774L236.972 373.823ZM226.562 372.314C225.442 372.617 224.78 373.769 225.083 374.889C225.386 376.008 226.539 376.67 227.658 376.367L226.562 372.314ZM218.326 378.856C219.448 378.561 220.118 377.414 219.824 376.292C219.53 375.171 218.382 374.5 217.26 374.794L218.326 378.856ZM207.946 377.19C206.821 377.472 206.139 378.613 206.421 379.737C206.704 380.862 207.845 381.544 208.969 381.262L207.946 377.19ZM199.563 383.559C200.691 383.292 201.389 382.161 201.122 381.033C200.856 379.905 199.724 379.206 198.596 379.473L199.563 383.559ZM189.22 381.605C188.087 381.852 187.369 382.97 187.615 384.103C187.861 385.236 188.98 385.955 190.113 385.708L189.22 381.605ZM180.616 387.662C181.755 387.442 182.5 386.341 182.28 385.203C182.06 384.064 180.959 383.319 179.821 383.539L180.616 387.662ZM170.405 385.214C169.261 385.399 168.483 386.477 168.667 387.622C168.852 388.766 169.93 389.544 171.075 389.359L170.405 385.214ZM161.43 390.728C162.581 390.589 163.402 389.544 163.264 388.392C163.125 387.241 162.08 386.42 160.928 386.559L161.43 390.728ZM151.434 387.452C150.277 387.528 149.4 388.527 149.475 389.684C149.551 390.841 150.55 391.718 151.707 391.643L151.434 387.452ZM141.919 391.925C143.078 391.939 144.029 391.01 144.043 389.85C144.057 388.691 143.128 387.74 141.969 387.726L141.919 391.925ZM132.56 386.924C131.414 386.746 130.341 387.531 130.163 388.676C129.985 389.822 130.769 390.895 131.915 391.073L132.56 386.924ZM122.156 389.064C123.278 389.357 124.425 388.685 124.718 387.564C125.011 386.442 124.34 385.295 123.218 385.001L122.156 389.064ZM114.143 382.023C113.068 381.59 111.844 382.11 111.411 383.185C110.977 384.261 111.497 385.484 112.573 385.918L114.143 382.023ZM103.513 381.44C104.508 382.036 105.797 381.712 106.393 380.717C106.988 379.722 106.665 378.433 105.67 377.837L103.513 381.44ZM98.0749 372.317C97.2042 371.551 95.8777 371.636 95.1119 372.506C94.3461 373.377 94.4311 374.704 95.3017 375.469L98.0749 372.317ZM88.3428 368.034C89.0489 368.953 90.3669 369.126 91.2866 368.42C92.2063 367.714 92.3795 366.396 91.6734 365.477L88.3428 368.034ZM86.6873 357.522C86.1641 356.487 84.9012 356.072 83.8664 356.596C82.8316 357.119 82.4169 358.382 82.9401 359.417L86.6873 357.522ZM79.1421 350.031C79.4909 351.137 80.67 351.75 81.7758 351.402C82.8817 351.053 83.4954 349.874 83.1466 348.768L79.1421 350.031ZM80.9195 339.522C80.7207 338.38 79.6336 337.615 78.4912 337.814C77.3489 338.012 76.5839 339.099 76.7826 340.242L80.9195 339.522ZM75.6125 330.293C75.6889 331.45 76.6888 332.326 77.8457 332.25C79.0028 332.173 79.8787 331.173 79.8023 330.016L75.6125 330.293ZM79.5886 320.398C79.6098 319.239 78.687 318.282 77.5277 318.261C76.3684 318.24 75.4115 319.162 75.3904 320.322L79.5886 320.398ZM75.915 310.396C75.8167 311.552 76.6736 312.568 77.829 312.666C78.9843 312.764 80.0006 311.907 80.0989 310.752L75.915 310.396ZM81.184 301.123C81.3428 299.975 80.5405 298.915 79.3919 298.756C78.2433 298.597 77.1834 299.399 77.0246 300.548L81.184 301.123ZM78.5853 290.788C78.3802 291.93 79.1392 293.021 80.2804 293.226C81.4216 293.431 82.513 292.672 82.7181 291.531L78.5853 290.788ZM84.5823 281.987C84.8192 280.852 84.091 279.739 82.956 279.502C81.8209 279.266 80.7087 279.994 80.4719 281.129L84.5823 281.987ZM82.5536 271.502C82.3048 272.634 83.0212 273.754 84.1537 274.003C85.2862 274.252 86.406 273.535 86.6548 272.403L82.5536 271.502ZM405.98 498.685C406.998 497.519 408.06 496.141 409.14 494.578L405.686 492.191C404.677 493.65 403.712 494.898 402.818 495.922L405.98 498.685ZM414.357 485.805C415.814 482.997 417.253 479.904 418.615 476.575L414.729 474.985C413.414 478.2 412.028 481.178 410.63 483.87L414.357 485.805ZM422.06 467.057C423.051 463.93 423.951 460.665 424.724 457.291L420.632 456.353C419.885 459.611 419.015 462.766 418.057 465.788L422.06 467.057ZM426.568 447.333C427.032 444.03 427.358 440.657 427.517 437.236L423.323 437.041C423.17 440.324 422.857 443.567 422.409 446.748L426.568 447.333ZM427.477 427.074C427.289 423.715 426.92 420.334 426.344 416.952L422.205 417.657C422.753 420.875 423.105 424.099 423.285 427.308L427.477 427.074ZM424.025 407.011C423.061 403.794 421.885 400.597 420.473 397.439L416.64 399.153C417.975 402.14 419.089 405.167 420.003 408.216L424.025 407.011ZM415.701 388.41C413.952 385.571 411.984 382.783 409.78 380.063L406.518 382.707C408.608 385.286 410.471 387.926 412.126 390.612L415.701 388.41ZM402.908 372.568C400.541 370.276 397.977 368.05 395.204 365.9L392.631 369.218C395.285 371.276 397.733 373.402 399.987 375.585L402.908 372.568ZM386.877 360.082C384.123 358.346 381.203 356.675 378.111 355.074L376.181 358.803C379.171 360.351 381.988 361.964 384.639 363.634L386.877 360.082ZM368.981 350.777C367.438 350.117 365.859 349.473 364.242 348.846L362.722 352.76C364.296 353.371 365.832 353.997 367.33 354.638L368.981 350.777ZM364.242 348.846C362.709 348.251 361.078 347.76 359.361 347.364L358.418 351.456C359.963 351.812 361.399 352.246 362.722 352.76L364.242 348.846ZM349.431 346.094C346.33 345.953 343.035 345.997 339.574 346.2L339.819 350.392C343.164 350.195 346.31 350.155 349.24 350.289L349.431 346.094ZM329.857 347.114C326.741 347.504 323.519 347.99 320.204 348.557L320.913 352.696C324.175 352.137 327.334 351.661 330.379 351.28L329.857 347.114ZM310.668 350.356C307.575 350.988 304.417 351.676 301.204 352.411L302.14 356.504C305.326 355.776 308.451 355.095 311.508 354.47L310.668 350.356ZM291.779 354.655C288.689 355.417 285.558 356.213 282.393 357.035L283.449 361.099C286.6 360.281 289.714 359.489 292.785 358.732L291.779 354.655ZM273.055 359.508C269.972 360.337 266.866 361.184 263.742 362.043L264.855 366.091C267.973 365.234 271.072 364.389 274.146 363.562L273.055 359.508ZM254.442 364.615C251.353 365.473 248.25 366.338 245.148 367.2L246.273 371.246C249.379 370.382 252.476 369.519 255.566 368.661L254.442 364.615ZM235.856 369.774C232.753 370.63 229.653 371.478 226.562 372.314L227.658 376.367C230.757 375.529 233.864 374.679 236.972 373.823L235.856 369.774ZM217.26 374.794C214.137 375.614 211.03 376.415 207.946 377.19L208.969 381.262C212.069 380.483 215.191 379.679 218.326 378.856L217.26 374.794ZM198.596 379.473C195.435 380.221 192.307 380.934 189.22 381.605L190.113 385.708C193.227 385.031 196.38 384.312 199.563 383.559L198.596 379.473ZM179.821 383.539C176.623 384.156 173.481 384.717 170.405 385.214L171.075 389.359C174.198 388.855 177.381 388.286 180.616 387.662L179.821 383.539ZM160.928 386.559C157.663 386.952 154.494 387.253 151.434 387.452L151.707 391.643C154.853 391.438 158.099 391.128 161.43 390.728L160.928 386.559ZM141.969 387.726C140.365 387.707 138.802 387.651 137.283 387.556L137.02 391.746C138.616 391.846 140.249 391.905 141.919 391.925L141.969 387.726ZM137.383 387.564C135.73 387.381 134.123 387.167 132.56 386.924L131.915 391.073C133.54 391.326 135.208 391.548 136.92 391.738L137.383 387.564ZM123.218 385.001C119.964 384.151 116.944 383.153 114.143 382.023L112.573 385.918C115.549 387.118 118.74 388.171 122.156 389.064L123.218 385.001ZM105.67 377.837C102.862 376.157 100.339 374.308 98.0749 372.317L95.3017 375.469C97.7614 377.633 100.491 379.631 103.513 381.44L105.67 377.837ZM91.6734 365.477C89.7479 362.969 88.0959 360.308 86.6873 357.522L82.9401 359.417C84.4599 362.422 86.2493 365.307 88.3428 368.034L91.6734 365.477ZM83.1466 348.768C82.2018 345.772 81.4689 342.681 80.9195 339.522L76.7826 340.242C77.3607 343.565 78.1356 346.839 79.1421 350.031L83.1466 348.768ZM79.8023 330.016C79.592 326.832 79.5301 323.616 79.5886 320.398L75.3904 320.322C75.3299 323.644 75.3935 326.978 75.6125 330.293L79.8023 330.016ZM80.0989 310.752C80.3754 307.501 80.7472 304.281 81.184 301.123L77.0246 300.548C76.5791 303.769 76.1986 307.063 75.915 310.396L80.0989 310.752ZM82.7181 291.531C83.3133 288.219 83.9485 285.024 84.5823 281.987L80.4719 281.129C79.8325 284.193 79.1892 287.427 78.5853 290.788L82.7181 291.531ZM86.6548 272.403C87.0229 270.727 87.372 269.132 87.6901 267.632L83.5824 266.761C83.2679 268.244 82.9217 269.826 82.5536 271.502L86.6548 272.403Z"
-    fill="black"
-   />
-   <path
-    d="M580.465 496.162C579.598 496.933 579.521 498.26 580.292 499.126C581.063 499.992 582.39 500.069 583.256 499.298L580.465 496.162ZM742.971 150.901C744.123 151.035 745.165 150.209 745.298 149.057L747.473 130.288C747.606 129.136 746.781 128.094 745.629 127.96C744.477 127.827 743.435 128.653 743.302 129.804L741.369 146.489L724.685 144.556C723.533 144.422 722.491 145.248 722.358 146.4C722.224 147.551 723.05 148.593 724.202 148.727L742.971 150.901ZM611.777 276.817L610.068 278.037L611.777 276.817ZM741.677 143.57C740.941 142.675 739.618 142.546 738.722 143.282C737.826 144.018 737.697 145.341 738.434 146.237L741.677 143.57ZM732.172 138.903C732.941 139.77 734.268 139.849 735.135 139.079C736.003 138.31 736.082 136.983 735.312 136.116L732.172 138.903ZM728.54 128.798C727.734 127.964 726.405 127.942 725.571 128.747C724.737 129.553 724.715 130.882 725.52 131.716L728.54 128.798ZM718.667 124.941C719.511 125.737 720.84 125.698 721.635 124.854C722.431 124.01 722.392 122.682 721.548 121.886L718.667 124.941ZM714.189 115.296C713.305 114.546 711.98 114.654 711.23 115.538C710.48 116.422 710.588 117.747 711.472 118.498L714.189 115.296ZM703.873 112.42C704.8 113.116 706.116 112.929 706.813 112.002C707.509 111.075 707.322 109.759 706.395 109.063L703.873 112.42ZM698.267 103.362C697.296 102.729 695.995 103.004 695.362 103.975C694.729 104.947 695.003 106.247 695.975 106.88L698.267 103.362ZM687.745 101.942C688.761 102.5 690.038 102.128 690.595 101.111C691.153 100.094 690.781 98.8183 689.764 98.2607L687.745 101.942ZM680.856 93.855C679.795 93.3874 678.556 93.8684 678.089 94.9295C677.621 95.9905 678.102 97.2298 679.163 97.6974L680.856 93.855ZM670.235 94.2686C671.337 94.629 672.523 94.0277 672.883 92.9256C673.244 91.8236 672.642 90.638 671.54 90.2776L670.235 94.2686ZM661.852 87.6927C660.716 87.4589 659.606 88.1899 659.372 89.3256C659.138 90.4613 659.869 91.5715 661.005 91.8054L661.852 87.6927ZM651.564 90.4692C652.72 90.5569 653.729 89.6907 653.816 88.5345C653.904 87.3783 653.038 86.3699 651.882 86.2823L651.564 90.4692ZM641.785 86.2189C640.628 86.2944 639.751 87.2936 639.827 88.4506C639.902 89.6077 640.901 90.4845 642.058 90.409L641.785 86.2189ZM632.676 91.7314C633.809 91.4825 634.525 90.3626 634.276 89.2302C634.027 88.0977 632.907 87.3814 631.775 87.6303L632.676 91.7314ZM622.099 90.5654C621.02 90.9881 620.487 92.206 620.91 93.2857C621.332 94.3655 622.55 94.8981 623.63 94.4754L622.099 90.5654ZM615.126 98.6009C616.127 98.015 616.463 96.7289 615.877 95.7283C615.291 94.7278 614.005 94.3916 613.005 94.9775L615.126 98.6009ZM604.699 100.733C603.798 101.463 603.66 102.785 604.39 103.686C605.12 104.587 606.442 104.725 607.343 103.995L604.699 100.733ZM600.415 110.489C601.203 109.639 601.153 108.311 600.302 107.522C599.452 106.734 598.124 106.785 597.335 107.635L600.415 110.489ZM591.007 115.447C590.337 116.393 590.561 117.704 591.507 118.374C592.453 119.044 593.763 118.82 594.434 117.874L591.007 115.447ZM589.412 125.972C589.963 124.952 589.584 123.678 588.564 123.127C587.544 122.575 586.27 122.955 585.718 123.975L589.412 125.972ZM581.452 133.038C581.016 134.112 581.533 135.337 582.607 135.773C583.682 136.209 584.906 135.692 585.343 134.618L581.452 133.038ZM582.198 143.664C582.524 142.551 581.886 141.385 580.773 141.059C579.66 140.733 578.494 141.371 578.168 142.484L582.198 143.664ZM575.816 152.189C575.595 153.328 576.339 154.43 577.477 154.651C578.615 154.872 579.717 154.129 579.938 152.99L575.816 152.189ZM578.517 162.498C578.639 161.345 577.804 160.311 576.651 160.188C575.498 160.065 574.464 160.901 574.341 162.054L578.517 162.498ZM573.688 171.997C573.657 173.157 574.573 174.121 575.732 174.151C576.891 174.181 577.855 173.266 577.885 172.107L573.688 171.997ZM577.997 181.755C577.941 180.597 576.957 179.703 575.798 179.76C574.64 179.816 573.747 180.8 573.803 181.958L577.997 181.755ZM574.638 191.888C574.776 193.039 575.82 193.861 576.972 193.724C578.123 193.587 578.945 192.542 578.808 191.391L574.638 191.888ZM580.265 200.921C580.053 199.781 578.957 199.029 577.817 199.241C576.677 199.454 575.925 200.55 576.137 201.69L580.265 200.921ZM578.254 211.357C578.536 212.482 579.677 213.164 580.802 212.882C581.926 212.599 582.609 211.459 582.326 210.334L578.254 211.357ZM584.957 219.629C584.609 218.523 583.43 217.909 582.324 218.257C581.218 218.605 580.603 219.783 580.951 220.889L584.957 219.629ZM584.196 230.274C584.605 231.359 585.816 231.907 586.901 231.498C587.986 231.088 588.534 229.877 588.125 228.792L584.196 230.274ZM591.797 237.795C591.331 236.733 590.092 236.25 589.031 236.716C587.969 237.182 587.486 238.42 587.952 239.482L591.797 237.795ZM592.132 248.393C592.65 249.431 593.911 249.851 594.949 249.333C595.986 248.815 596.407 247.554 595.888 246.517L592.132 248.393ZM600.425 255.078C599.858 254.066 598.579 253.706 597.567 254.273C596.556 254.84 596.195 256.119 596.762 257.131L600.425 255.078ZM601.828 265.7C602.441 266.685 603.735 266.986 604.72 266.374C605.705 265.761 606.006 264.467 605.394 263.482L601.828 265.7ZM610.661 271.556C610.007 270.598 608.701 270.352 607.743 271.006C606.786 271.66 606.54 272.967 607.194 273.924L610.661 271.556ZM612.943 282.123C613.603 283.076 614.911 283.314 615.864 282.654C616.817 281.994 617.055 280.686 616.395 279.732L612.943 282.123ZM621.852 287.874C621.221 286.9 619.921 286.623 618.948 287.253C617.975 287.884 617.697 289.184 618.328 290.157L621.852 287.874ZM623.545 298.516C624.141 299.51 625.431 299.833 626.425 299.236C627.419 298.64 627.742 297.35 627.145 296.356L623.545 298.516ZM632.09 304.953C631.531 303.937 630.254 303.566 629.238 304.125C628.222 304.684 627.852 305.96 628.41 306.976L632.09 304.953ZM632.91 315.568C633.426 316.606 634.687 317.029 635.725 316.513C636.763 315.997 637.186 314.737 636.67 313.699L632.91 315.568ZM640.876 322.659C640.408 321.598 639.168 321.117 638.107 321.585C637.047 322.054 636.566 323.293 637.034 324.354L640.876 322.659ZM640.751 333.387C641.164 334.471 642.377 335.014 643.461 334.601C644.544 334.188 645.087 332.975 644.674 331.891L640.751 333.387ZM647.964 341.299C647.613 340.194 646.433 339.583 645.328 339.934C644.223 340.285 643.611 341.465 643.963 342.57L647.964 341.299ZM646.615 351.903C646.896 353.028 648.036 353.712 649.161 353.431C650.286 353.149 650.97 352.009 650.688 350.884L646.615 351.903ZM652.78 360.635C652.576 359.494 651.486 358.733 650.345 358.937C649.203 359.14 648.443 360.231 648.646 361.372L652.78 360.635ZM649.99 370.947C650.107 372.101 651.138 372.94 652.291 372.823C653.445 372.706 654.285 371.675 654.167 370.522L649.99 370.947ZM654.782 380.495C654.758 379.335 653.798 378.415 652.639 378.439C651.48 378.463 650.56 379.422 650.583 380.581L654.782 380.495ZM650.374 390.216C650.299 391.373 651.176 392.372 652.333 392.447C653.49 392.522 654.489 391.645 654.564 390.488L650.374 390.216ZM653.474 400.432C653.651 399.286 652.867 398.213 651.721 398.035C650.575 397.857 649.502 398.642 649.324 399.788L653.474 400.432ZM647.411 409.257C647.13 410.382 647.815 411.521 648.939 411.802C650.064 412.083 651.204 411.399 651.485 410.274L647.411 409.257ZM648.619 419.88C649 418.785 648.421 417.588 647.326 417.207C646.231 416.826 645.035 417.405 644.654 418.5L648.619 419.88ZM641.091 427.456C640.617 428.514 641.089 429.756 642.147 430.231C643.205 430.705 644.447 430.233 644.922 429.175L641.091 427.456ZM640.447 438.129C641.007 437.114 640.638 435.837 639.623 435.277C638.608 434.716 637.331 435.085 636.77 436.1L640.447 438.129ZM631.768 444.384C631.132 445.353 631.402 446.655 632.371 447.291C633.34 447.928 634.642 447.658 635.278 446.689L631.768 444.384ZM629.514 454.822C630.217 453.9 630.038 452.582 629.116 451.88C628.193 451.177 626.876 451.355 626.174 452.278L629.514 454.822ZM620.04 459.817C619.281 460.694 619.376 462.019 620.252 462.779C621.128 463.538 622.454 463.443 623.214 462.566L620.04 459.817ZM616.498 469.89C617.304 469.057 617.283 467.728 616.45 466.922C615.616 466.115 614.287 466.137 613.481 466.97L616.498 469.89ZM606.529 473.81C605.684 474.604 605.642 475.933 606.436 476.778C607.23 477.623 608.559 477.664 609.404 476.87L606.529 473.81ZM602.036 483.522C602.911 482.76 603.002 481.434 602.241 480.56C601.48 479.685 600.154 479.593 599.279 480.354L602.036 483.522ZM591.808 486.68C590.915 487.419 590.791 488.743 591.53 489.636C592.27 490.529 593.594 490.653 594.487 489.914L591.808 486.68ZM586.902 496.174C587.792 495.431 587.91 494.107 587.167 493.217C586.424 492.327 585.1 492.208 584.21 492.951L586.902 496.174ZM744.858 147.512C743.798 146.174 742.738 144.86 741.677 143.57L738.434 146.237C739.478 147.507 740.523 148.801 741.567 150.12L744.858 147.512ZM735.312 136.116C733.054 133.571 730.796 131.132 728.54 128.798L725.52 131.716C727.734 134.006 729.951 136.401 732.172 138.903L735.312 136.116ZM721.548 121.886C719.09 119.569 716.636 117.373 714.189 115.296L711.472 118.498C713.862 120.525 716.261 122.672 718.667 124.941L721.548 121.886ZM706.395 109.063C703.674 107.018 700.964 105.119 698.267 103.362L695.975 106.88C698.59 108.584 701.224 110.429 703.873 112.42L706.395 109.063ZM689.764 98.2607C686.772 96.6198 683.801 95.1529 680.856 93.855L679.163 97.6974C681.993 98.9446 684.855 100.357 687.745 101.942L689.764 98.2607ZM671.54 90.2776C668.27 89.2081 665.038 88.3489 661.852 87.6927L661.005 91.8054C664.031 92.4286 667.11 93.2466 670.235 94.2686L671.54 90.2776ZM651.882 86.2823C648.446 86.0217 645.077 86.0042 641.785 86.2189L642.058 90.409C645.15 90.2073 648.321 90.2233 651.564 90.4692L651.882 86.2823ZM631.775 87.6303C628.445 88.3623 625.216 89.3454 622.099 90.5654L623.63 94.4754C626.542 93.3354 629.561 92.4162 632.676 91.7314L631.775 87.6303ZM613.005 94.9775C610.103 96.6763 607.331 98.6 604.699 100.733L607.343 103.995C609.814 101.993 612.412 100.191 615.126 98.6009L613.005 94.9775ZM597.335 107.635C595.083 110.065 592.97 112.674 591.007 115.447L594.434 117.874C596.294 115.246 598.291 112.781 600.415 110.489L597.335 107.635ZM585.718 123.975C584.154 126.866 582.73 129.891 581.452 133.038L585.343 134.618C586.564 131.609 587.923 128.724 589.412 125.972L585.718 123.975ZM578.168 142.484C577.249 145.625 576.463 148.863 575.816 152.189L579.938 152.99C580.56 149.789 581.316 146.677 582.198 143.664L578.168 142.484ZM574.341 162.054C573.996 165.301 573.776 168.618 573.688 171.997L577.885 172.107C577.971 168.838 578.183 165.632 578.517 162.498L574.341 162.054ZM573.803 181.958C573.962 185.221 574.239 188.533 574.638 191.888L578.808 191.391C578.42 188.132 578.151 184.918 577.997 181.755L573.803 181.958ZM576.137 201.69C576.732 204.881 577.437 208.105 578.254 211.357L582.326 210.334C581.53 207.165 580.844 204.026 580.265 200.921L576.137 201.69ZM580.951 220.889C581.929 223.998 583.01 227.127 584.196 230.274L588.125 228.792C586.966 225.718 585.911 222.662 584.957 219.629L580.951 220.889ZM587.952 239.482C589.25 242.442 590.643 245.413 592.132 248.393L595.888 246.517C594.43 243.598 593.067 240.69 591.797 237.795L587.952 239.482ZM596.762 257.131C598.361 259.983 600.049 262.84 601.828 265.7L605.394 263.482C603.648 260.676 601.992 257.873 600.425 255.078L596.762 257.131ZM607.194 273.924C608.131 275.295 609.089 276.666 610.068 278.037L613.485 275.596C612.522 274.249 611.581 272.902 610.661 271.556L607.194 273.924ZM610.068 278.037C611.047 279.407 612.005 280.769 612.943 282.123L616.395 279.732C615.445 278.361 614.475 276.982 613.485 275.596L610.068 278.037ZM618.328 290.157C620.155 292.977 621.893 295.764 623.545 298.516L627.145 296.356C625.468 293.561 623.705 290.734 621.852 287.874L618.328 290.157ZM628.41 306.976C630.007 309.879 631.506 312.743 632.91 315.568L636.67 313.699C635.239 310.821 633.713 307.906 632.09 304.953L628.41 306.976ZM637.034 324.354C638.384 327.413 639.621 330.424 640.751 333.387L644.674 331.891C643.518 328.859 642.254 325.782 640.876 322.659L637.034 324.354ZM643.963 342.57C644.969 345.738 645.851 348.849 646.615 351.903L650.688 350.884C649.903 347.743 648.997 344.548 647.964 341.299L643.963 342.57ZM648.646 361.372C649.227 364.631 649.672 367.823 649.99 370.947L654.167 370.522C653.838 367.291 653.379 363.996 652.78 360.635L648.646 361.372ZM650.583 380.581C650.652 383.871 650.578 387.082 650.374 390.216L654.564 390.488C654.776 387.234 654.852 383.904 654.782 380.495L650.583 380.581ZM649.324 399.788C648.821 403.032 648.178 406.187 647.411 409.257L651.485 410.274C652.282 407.083 652.95 403.803 653.474 400.432L649.324 399.788ZM644.654 418.5C643.583 421.579 642.389 424.563 641.091 427.456L644.922 429.175C646.268 426.175 647.507 423.078 648.619 419.88L644.654 418.5ZM636.77 436.1C635.191 438.961 633.517 441.721 631.768 444.384L635.278 446.689C637.082 443.941 638.813 441.089 640.447 438.129L636.77 436.1ZM626.174 452.278C624.178 454.898 622.125 457.41 620.04 459.817L623.214 462.566C625.352 460.098 627.461 457.517 629.514 454.822L626.174 452.278ZM613.481 466.97C611.157 469.371 608.828 471.65 606.529 473.81L609.404 476.87C611.745 474.67 614.122 472.345 616.498 469.89L613.481 466.97ZM599.279 480.354C596.674 482.622 594.166 484.727 591.808 486.68L594.487 489.914C596.858 487.949 599.396 485.82 602.036 483.522L599.279 480.354ZM584.21 492.951C582.849 494.088 581.591 495.159 580.465 496.162L583.256 499.298C584.336 498.337 585.556 497.298 586.902 496.174L584.21 492.951Z"
-    fill="black"
-   />
-   <rect
-    x="798.315"
-    y="446.884"
-    width="171.375"
-    height="39.3156"
-    rx="4.41052"
-    transform="rotate(-51.2011 798.315 446.884)"
-    fill="white"
-   />
-   <rect
-    x="810.238"
-    y="445.051"
-    width="168.288"
-    height="36.2282"
-    rx="2.86684"
-    transform="rotate(-49.6275 810.238 445.051)"
-    stroke="black"
-    strokeWidth="3.08737"
-   />
-   <line
-    x1="824.001"
-    y1="432.268"
-    x2="830.176"
-    y2="437.672"
-    stroke="black"
-    strokeWidth="4.41052"
-    strokeLinecap="round"
-   />
-   <line
-    x1="861.794"
-    y1="387.22"
-    x2="867.969"
-    y2="392.625"
-    stroke="black"
-    strokeWidth="4.41052"
-    strokeLinecap="round"
-   />
-   <line
-    x1="842.881"
-    y1="409.939"
-    x2="849.056"
-    y2="415.344"
-    stroke="black"
-    strokeWidth="4.41052"
-    strokeLinecap="round"
-   />
-   <line
-    x1="880.673"
-    y1="364.892"
-    x2="886.848"
-    y2="370.296"
-    stroke="black"
-    strokeWidth="4.41052"
-    strokeLinecap="round"
-   />
-   <line
-    x1="900.01"
-    y1="341.817"
-    x2="906.185"
-    y2="347.221"
-    stroke="black"
-    strokeWidth="4.41052"
-    strokeLinecap="round"
-   />
-   <line
-    x1="832.907"
-    y1="420.467"
-    x2="839.083"
-    y2="425.871"
-    stroke="black"
-    strokeWidth="4.41052"
-    strokeLinecap="round"
-   />
-   <line
-    x1="870.7"
-    y1="375.419"
-    x2="876.875"
-    y2="380.824"
-    stroke="black"
-    strokeWidth="4.41052"
-    strokeLinecap="round"
-   />
-   <line
-    x1="851.785"
-    y1="398.138"
-    x2="857.961"
-    y2="403.542"
-    stroke="black"
-    strokeWidth="4.41052"
-    strokeLinecap="round"
-   />
-   <line
-    x1="889.579"
-    y1="353.09"
-    x2="895.754"
-    y2="358.495"
-    stroke="black"
-    strokeWidth="4.41052"
-    strokeLinecap="round"
-   />
-   <line
-    x1="908.917"
-    y1="330.016"
-    x2="915.092"
-    y2="335.42"
-    stroke="black"
-    strokeWidth="4.41052"
-    strokeLinecap="round"
-   />
-   <path
-    d="M438.122 100.843C439.11 102.327 440.017 103.478 439.687 105.169C439.127 105.391 438.484 105.766 437.792 105.89C434.71 106.487 431.628 107.152 428.496 107.499C426.683 107.707 424.87 107.97 423.024 108.068C420.931 108.178 418.953 108.65 417.009 109.357C415.69 108.47 416.053 107.194 415.954 106.043C415.888 105.044 415.772 104.018 416.959 103.297C417.981 102.687 418.97 102.923 420.074 103.075C421.343 103.256 422.678 103.117 423.98 103.075C425.942 103.02 427.903 103.006 429.848 102.853C431.891 102.687 433.869 102.229 435.781 101.55C436.473 101.287 437.199 101.106 438.122 100.843Z"
-    fill="black"
-   />
-   <path
-    d="M220.78 424.314C220.059 424.41 219.442 424.381 218.915 424.572C217.61 425.066 216.323 425.554 214.874 425.496C214.818 425.494 214.73 425.482 214.704 425.514C214.056 426.43 212.999 426.087 212.092 426.338C211.541 425.306 211.002 424.28 211.496 423.045C212.354 422.581 213.335 422.456 214.339 422.368C215.984 422.223 217.556 421.797 219.073 421.171C219.172 421.135 219.239 421.012 219.397 420.856C219.031 418.711 217.26 414.798 215.992 412.892C216.244 412.636 216.447 412.321 216.73 412.174C217.192 411.93 217.702 411.767 218.202 411.63C219.301 411.342 220.579 413.041 220.876 414.352C221.052 415.147 221.588 416.961 221.718 417.76C221.851 418.537 222.301 419.111 222.644 419.801C225.099 418.956 227.501 418.273 229.455 416.587C230.635 417.501 230.838 418.781 230.934 420.077C230.955 420.367 230.779 420.705 230.601 420.935C229.939 421.812 229.017 422.047 227.979 422.093C226.439 422.167 225.058 422.733 223.72 423.547C223.811 424.878 224.637 425.899 224.869 427.08C225.095 428.219 226.45 430.055 226.401 431.337C225.706 431.787 225.02 432.366 223.99 432.2C223.296 431.648 223.147 430.723 222.868 429.9C222.487 428.795 221.798 426.663 221.417 425.558C221.282 425.211 221.062 424.861 220.78 424.314Z"
-    fill="black"
-   />
-  </svg>
- );
-}
