--- src/components/svg/PhysiqueIllustration.jsx
@@ -1,34 +1,0 @@
-export default function PhysiqueIllustration() {
- return (
-  <svg
-   width="269"
-   height="323"
-   viewBox="0 0 269 323"
-   fill="none"
-   xmlns="http://www.w3.org/2000/svg"
-  >
-   <rect
-    x="2"
-    y="2"
-    width="265"
-    height="319"
-    rx="7"
-    fill="#DAABB4"
-    stroke="black"
-    strokeWidth="4"
-   />
-   <path
-    d="M158.374 216H137.935L129.81 194.862H92.6123L84.9316 216H65L101.245 122.943H121.113L158.374 216ZM123.779 179.184L110.957 144.652L98.3887 179.184H123.779ZM166.634 216V122.943H184.471V216H166.634Z"
-    fill="black"
-   />
-   <path
-    d="M25.76 63V40.84L21.16 41.92V38L27.92 35H31.08V63H25.76ZM44.6691 63.48C42.8291 63.48 41.1624 63.16 39.6691 62.52C38.1757 61.8533 36.9757 60.84 36.0691 59.48C35.1624 58.12 34.6824 56.4133 34.6291 54.36H39.6691C39.6957 55.72 40.1357 56.8667 40.9891 57.8C41.8691 58.7067 43.0957 59.16 44.6691 59.16C46.1624 59.16 47.3091 58.7467 48.1091 57.92C48.9091 57.0933 49.3091 56.0533 49.3091 54.8C49.3091 53.3333 48.7757 52.2267 47.7091 51.48C46.6691 50.7067 45.3224 50.32 43.6691 50.32H41.5891V46.12H43.7091C45.0691 46.12 46.2024 45.8 47.1091 45.16C48.0157 44.52 48.4691 43.5733 48.4691 42.32C48.4691 41.28 48.1224 40.4533 47.4291 39.84C46.7624 39.2 45.8291 38.88 44.6291 38.88C43.3224 38.88 42.2957 39.2667 41.5491 40.04C40.8291 40.8133 40.4291 41.76 40.3491 42.88H35.3491C35.4557 40.2933 36.3491 38.2533 38.0291 36.76C39.7357 35.2667 41.9357 34.52 44.6291 34.52C46.5491 34.52 48.1624 34.8667 49.4691 35.56C50.8024 36.2267 51.8024 37.12 52.4691 38.24C53.1624 39.36 53.5091 40.6 53.5091 41.96C53.5091 43.5333 53.0691 44.8667 52.1891 45.96C51.3357 47.0267 50.2691 47.7467 48.9891 48.12C50.5624 48.44 51.8424 49.2133 52.8291 50.44C53.8157 51.64 54.3091 53.16 54.3091 55C54.3091 56.5467 53.9357 57.96 53.1891 59.24C52.4424 60.52 51.3491 61.5467 49.9091 62.32C48.4957 63.0933 46.7491 63.48 44.6691 63.48Z"
-    fill="black"
-   />
-   <path
-    d="M59.1719 287.441C57.8698 288.548 56.6133 289.329 55.4023 289.785C54.2044 290.241 52.9154 290.469 51.5352 290.469C49.2565 290.469 47.5052 289.915 46.2812 288.809C45.0573 287.689 44.4453 286.263 44.4453 284.531C44.4453 283.516 44.6732 282.591 45.1289 281.758C45.5977 280.911 46.2031 280.234 46.9453 279.727C47.7005 279.219 48.5469 278.835 49.4844 278.574C50.1745 278.392 51.2161 278.216 52.6094 278.047C55.4479 277.708 57.5378 277.305 58.8789 276.836C58.8919 276.354 58.8984 276.048 58.8984 275.918C58.8984 274.486 58.5664 273.477 57.9023 272.891C57.0039 272.096 55.6693 271.699 53.8984 271.699C52.2448 271.699 51.0208 271.992 50.2266 272.578C49.4453 273.151 48.8659 274.173 48.4883 275.645L45.0508 275.176C45.3633 273.704 45.8776 272.52 46.5938 271.621C47.3099 270.71 48.3451 270.013 49.6992 269.531C51.0534 269.036 52.6224 268.789 54.4062 268.789C56.1771 268.789 57.6159 268.997 58.7227 269.414C59.8294 269.831 60.6432 270.358 61.1641 270.996C61.6849 271.621 62.0495 272.415 62.2578 273.379C62.375 273.978 62.4336 275.059 62.4336 276.621V281.309C62.4336 284.577 62.5052 286.647 62.6484 287.52C62.8047 288.379 63.1042 289.206 63.5469 290H59.875C59.5104 289.271 59.276 288.418 59.1719 287.441ZM58.8789 279.59C57.6029 280.111 55.6888 280.553 53.1367 280.918C51.6914 281.126 50.6693 281.361 50.0703 281.621C49.4714 281.882 49.0091 282.266 48.6836 282.773C48.3581 283.268 48.1953 283.822 48.1953 284.434C48.1953 285.371 48.5469 286.152 49.25 286.777C49.9661 287.402 51.0078 287.715 52.375 287.715C53.7292 287.715 54.9336 287.422 55.9883 286.836C57.043 286.237 57.8177 285.423 58.3125 284.395C58.6901 283.6 58.8789 282.428 58.8789 280.879V279.59ZM67.8242 290V261.367H71.3398V290H67.8242ZM90.4023 290V286.953C88.7878 289.297 86.5938 290.469 83.8203 290.469C82.5964 290.469 81.4505 290.234 80.3828 289.766C79.3281 289.297 78.5404 288.711 78.0195 288.008C77.5117 287.292 77.1536 286.419 76.9453 285.391C76.8021 284.701 76.7305 283.607 76.7305 282.109V269.258H80.2461V280.762C80.2461 282.598 80.3177 283.835 80.4609 284.473C80.6823 285.397 81.151 286.126 81.8672 286.66C82.5833 287.181 83.4688 287.441 84.5234 287.441C85.5781 287.441 86.5677 287.174 87.4922 286.641C88.4167 286.094 89.0677 285.358 89.4453 284.434C89.8359 283.496 90.0312 282.142 90.0312 280.371V269.258H93.5469V290H90.4023ZM99.0742 290V269.258H102.219V272.168C102.87 271.152 103.736 270.339 104.816 269.727C105.897 269.102 107.128 268.789 108.508 268.789C110.044 268.789 111.301 269.108 112.277 269.746C113.267 270.384 113.964 271.276 114.367 272.422C116.008 270 118.143 268.789 120.773 268.789C122.831 268.789 124.413 269.362 125.52 270.508C126.626 271.641 127.18 273.392 127.18 275.762V290H123.684V276.934C123.684 275.527 123.566 274.518 123.332 273.906C123.111 273.281 122.701 272.78 122.102 272.402C121.503 272.025 120.799 271.836 119.992 271.836C118.534 271.836 117.323 272.324 116.359 273.301C115.396 274.264 114.914 275.814 114.914 277.949V290H111.398V276.523C111.398 274.961 111.112 273.789 110.539 273.008C109.966 272.227 109.029 271.836 107.727 271.836C106.737 271.836 105.819 272.096 104.973 272.617C104.139 273.138 103.534 273.9 103.156 274.902C102.779 275.905 102.59 277.35 102.59 279.238V290H99.0742ZM132.414 265.41V261.367H135.93V265.41H132.414ZM132.414 290V269.258H135.93V290H132.414ZM141.301 290V269.258H144.465V272.207C145.988 269.928 148.189 268.789 151.066 268.789C152.316 268.789 153.462 269.017 154.504 269.473C155.559 269.915 156.346 270.501 156.867 271.23C157.388 271.96 157.753 272.826 157.961 273.828C158.091 274.479 158.156 275.618 158.156 277.246V290H154.641V277.383C154.641 275.951 154.504 274.883 154.23 274.18C153.957 273.464 153.469 272.897 152.766 272.48C152.076 272.051 151.262 271.836 150.324 271.836C148.827 271.836 147.531 272.311 146.438 273.262C145.357 274.212 144.816 276.016 144.816 278.672V290H141.301ZM163.586 265.41V261.367H167.102V265.41H163.586ZM163.586 290V269.258H167.102V290H163.586ZM186.066 290V286.953C184.452 289.297 182.258 290.469 179.484 290.469C178.26 290.469 177.115 290.234 176.047 289.766C174.992 289.297 174.204 288.711 173.684 288.008C173.176 287.292 172.818 286.419 172.609 285.391C172.466 284.701 172.395 283.607 172.395 282.109V269.258H175.91V280.762C175.91 282.598 175.982 283.835 176.125 284.473C176.346 285.397 176.815 286.126 177.531 286.66C178.247 287.181 179.133 287.441 180.188 287.441C181.242 287.441 182.232 287.174 183.156 286.641C184.081 286.094 184.732 285.358 185.109 284.434C185.5 283.496 185.695 282.142 185.695 280.371V269.258H189.211V290H186.066ZM194.738 290V269.258H197.883V272.168C198.534 271.152 199.4 270.339 200.48 269.727C201.561 269.102 202.792 268.789 204.172 268.789C205.708 268.789 206.965 269.108 207.941 269.746C208.931 270.384 209.628 271.276 210.031 272.422C211.672 270 213.807 268.789 216.438 268.789C218.495 268.789 220.077 269.362 221.184 270.508C222.29 271.641 222.844 273.392 222.844 275.762V290H219.348V276.934C219.348 275.527 219.23 274.518 218.996 273.906C218.775 273.281 218.365 272.78 217.766 272.402C217.167 272.025 216.464 271.836 215.656 271.836C214.198 271.836 212.987 272.324 212.023 273.301C211.06 274.264 210.578 275.814 210.578 277.949V290H207.062V276.523C207.062 274.961 206.776 273.789 206.203 273.008C205.63 272.227 204.693 271.836 203.391 271.836C202.401 271.836 201.483 272.096 200.637 272.617C199.803 273.138 199.198 273.9 198.82 274.902C198.443 275.905 198.254 277.35 198.254 279.238V290H194.738Z"
-    fill="black"
-   />
-  </svg>
- );
-}
